% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-evaluation.R
\name{APPA}
\alias{APPA}
\title{Average posterior probability of assignment (APPA)}
\usage{
APPA(object)
}
\arguments{
\item{object}{The model, of type \code{lcModel}.}
}
\value{
The APPA per cluster, as a \verb{numeric vector} of length \code{nClusters(object)}.
Empty clusters will output \code{NA}.
}
\description{
Computes the average posterior probability of assignment (APPA) for each cluster.
}
\references{
\insertRef{nagin2005group}{latrend}

\insertRef{klijn2017introducing}{latrend}

\insertRef{vandernest2020overview}{latrend}
}
\seealso{
\link{confusionMatrix} \link{OCC}
}
