% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{as.data.frame.lcMethods}
\alias{as.data.frame.lcMethods}
\title{Convert a list of lcMethod objects to a data.frame}
\usage{
\method{as.data.frame}{lcMethods}(x, ..., eval = FALSE, nullValue = NA, envir = NULL)
}
\arguments{
\item{x}{the \code{lcMethods} or \code{list} to be coerced to a \code{data.frame}.}

\item{...}{Additional arguments.}

\item{eval}{Whether to evaluate the arguments in order to replace expression if the resulting value is of a class specified in \code{evalClasses}.}

\item{nullValue}{Value to use to represent the \code{NULL} type. Must be of length 1.}

\item{envir}{The \code{environment} in which to evaluate the arguments. If \code{NULL}, the environment associated with the object is used. If not available, the \code{parent.frame()} is used.}
}
\value{
A \code{data.frame} with each row containing the argument values of a method object.
}
\description{
Converts a list of \code{lcMethod} objects to a \code{data.frame}.
}
\seealso{
Other lcMethod functions: 
\code{\link{[[,lcMethod-method}},
\code{\link{as.data.frame.lcMethod}()},
\code{\link{as.lcMethods}()},
\code{\link{as.list.lcMethod}()},
\code{\link{evaluate.lcMethod}()},
\code{\link{formula.lcMethod}()},
\code{\link{lcMethod-class}},
\code{\link{names,lcMethod-method}},
\code{\link{update.lcMethod}()}
}
\concept{lcMethod functions}
