% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{as.lcMethods}
\alias{as.lcMethods}
\title{Convert a list of lcMethod objects to a lcMethods list}
\usage{
as.lcMethods(x)
}
\arguments{
\item{x}{A \code{list} of \code{lcMethod} objects.}
}
\value{
A \code{lcMethods} object.
}
\description{
Convert a list of lcMethod objects to a lcMethods list
}
\seealso{
Other lcMethod functions: 
\code{\link{[[,lcMethod-method}},
\code{\link{as.data.frame.lcMethod}()},
\code{\link{as.data.frame.lcMethods}()},
\code{\link{as.list.lcMethod}()},
\code{\link{evaluate.lcMethod}()},
\code{\link{formula.lcMethod}()},
\code{\link{lcMethod-class}},
\code{\link{names,lcMethod-method}},
\code{\link{update.lcMethod}()}
}
\concept{lcMethod functions}
