% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{$,lcMethod-method}
\alias{$,lcMethod-method}
\title{Retrieve and evaluate a lcMethod argument by name}
\usage{
\S4method{$}{lcMethod}(x, name)
}
\arguments{
\item{x}{The \code{lcMethod} object.}

\item{name}{Name of the argument to retrieve.}
}
\value{
The argument evaluation result.
}
\description{
Retrieve and evaluate a lcMethod argument by name
}
\examples{
m <- lcMethodKML(nClusters = 3)
m$nClusters # 3
}
