% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{clusterNames}
\alias{clusterNames}
\title{Get the cluster names}
\usage{
clusterNames(object, factor = FALSE)
}
\arguments{
\item{object}{The \code{lcModel} object.}

\item{factor}{Whether to return the cluster names as a factor.}
}
\value{
A \code{character} of the cluster names.
}
\description{
Get the cluster names
}
\examples{
data(latrendData)
method <- lcMethodLMKM(Y ~ Time, id = "Id", time = "Time")
model <- latrend(method, latrendData)
clusterNames(model) # A, B
}
