% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{clusterProportions}
\alias{clusterProportions}
\alias{clusterProportions,lcModel-method}
\title{Proportional size of each cluster}
\usage{
\S4method{clusterProportions}{lcModel}(object, ...)
}
\arguments{
\item{object}{The \code{lcModel} to obtain the proportions from.}

\item{...}{Not used.}
}
\value{
A named \verb{numeric vector} of length \code{nClusters(object)} with the proportional size of each cluster.
}
\description{
Obtain the proportional size per cluster, with sizes between 0 and 1.
}
\details{
By default, the cluster proportions are computed from the average cluster weight from the posterior probabilities of the fitted data (as computed by the \code{\link[=postprob]{postprob()}} function).
}
\examples{
data(latrendData)
model <- latrend(lcMethodKML("Y", id = "Id", time = "Time"), latrendData)
clusterProportions(model)
}
\seealso{
\link{postprob}
}
