% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{coef.lcModel}
\alias{coef.lcModel}
\title{Extract lcModel coefficients}
\usage{
\method{coef}{lcModel}(object, ...)
}
\arguments{
\item{object}{The \code{lcModel} object.}

\item{...}{Additional arguments.}
}
\value{
A named \verb{numeric vector} with all coefficients, or a \code{matrix} with each column containing the cluster-specific coefficients. If \code{coef()} is not defined for the given model, an empty \verb{numeric vector} is returned.
}
\description{
Extract the coefficients of the \code{lcModel} object, if defined. The returned set of coefficients is dependent on the underlying type of \code{lcModel}.
}
\details{
The default implementation checks for the existence of a \code{coef()} function for the internal model, and returns the output, if available.
}
\seealso{
Other model-specific methods: 
\code{\link{clusterTrajectories}()},
\code{\link{converged}()},
\code{\link{deviance.lcModel}()},
\code{\link{df.residual.lcModel}()},
\code{\link{fitted.lcModel}()},
\code{\link{lcModel-class}},
\code{\link{logLik.lcModel}()},
\code{\link{model.frame.lcModel}()},
\code{\link{nobs.lcModel}()},
\code{\link{postprob}()},
\code{\link{predict.lcModel}()},
\code{\link{predictAssignments}()},
\code{\link{predictForCluster}()},
\code{\link{predictPostprob}()},
\code{\link{residuals.lcModel}()},
\code{\link{sigma.lcModel}()},
\code{\link{time.lcModel}()},
\code{\link{trajectories}()}
}
\concept{model-specific methods}
