% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latrend.R
\name{createTestDataFolds}
\alias{createTestDataFolds}
\title{Create all k test folds from the training data}
\usage{
createTestDataFolds(data, trainDataList, ...)
}
\arguments{
\item{data}{A \code{data.frame} representing the complete dataset.}

\item{trainDataList}{A \code{list} of \code{data.frame} representing each of the data training folds. These should be derived from \code{data}.}

\item{...}{Arguments passed to \link{createTestDataFold}.}
}
\description{
Create all k test folds from the training data
}
\examples{
data(latrendData)

if (require("caret")) {
  trainDataList <- createTrainDataFolds(latrendData, folds = 10, id = "Id")
  testDataList <- createTestDataFolds(latrendData, trainDataList)
}
}
\seealso{
Other validation methods: 
\code{\link{createTestDataFold}()},
\code{\link{createTrainDataFolds}()},
\code{\link{latrendBoot}()},
\code{\link{latrendCV}()},
\code{\link{lcModel-data-filters}}
}
\concept{validation methods}
