% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricsExternal.R
\name{defineExternalMetric}
\alias{defineExternalMetric}
\title{Define an external metric for lcModels}
\usage{
defineExternalMetric(
  name,
  fun,
  warnIfExists = getOption("latrend.warnMetricOverride", TRUE)
)
}
\arguments{
\item{name}{The name of the metric.}

\item{fun}{The function to compute the metric, accepting a lcModel object as input.}

\item{warnIfExists}{Whether to output a warning when the metric is already defined.}
}
\description{
Define an external metric for lcModels
}
\seealso{
Other metric functions: 
\code{\link{defineInternalMetric}()},
\code{\link{externalMetric}()},
\code{\link{getExternalMetricDefinition}()},
\code{\link{getExternalMetricNames}()},
\code{\link{getInternalMetricDefinition}()},
\code{\link{getInternalMetricNames}()},
\code{\link{metric}()}
}
\concept{metric functions}
