% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricsInternal.R
\name{.defineInternalDistanceMetrics}
\alias{.defineInternalDistanceMetrics}
\title{Define the distance metrics for multiple types at once}
\usage{
.defineInternalDistanceMetrics(
  name,
  type = c("traj", "fitted"),
  distanceFun,
  clusterAggregationFun = weighted.mean,
  assertNonEmpty = TRUE,
  assertNonSolitary = FALSE,
  assertNonIdentical = FALSE,
  ...
)
}
\description{
Define the distance metrics for multiple types at once
}
\keyword{internal}
