% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{.guessResponseVariable}
\alias{.guessResponseVariable}
\title{Guess the response variable}
\usage{
.guessResponseVariable(
  data,
  id = getOption("latrend.id"),
  time = getOption("latrend.time"),
  cluster = "Cluster"
)
}
\arguments{
\item{data}{A trajectories \code{data.frame}.}

\item{id}{Id variable(s) to exclude.}

\item{time}{Time variable(s) to exclude.}

\item{cluster}{Cluster variable(s) to exclude.}
}
\description{
Attempts to identify the response variable for the given trajectory data
}
\keyword{internal}
