% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latrend.R
\name{.trajSubset}
\alias{.trajSubset}
\title{Select trajectories}
\usage{
.trajSubset(data, ids, id)
}
\arguments{
\item{data}{The longitudinal dataset, a \code{data.frame}}

\item{ids}{The trajectory identifiers, \code{vector}}

\item{id}{The name of the id column}
}
\description{
Create a subset of the data with the given trajectories
}
\keyword{internal}
