% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{estimationTime}
\alias{estimationTime}
\title{Get the model estimation time}
\usage{
estimationTime(object, unit = "secs")
}
\arguments{
\item{object}{The \code{lcModel} object.}

\item{unit}{The time unit in which the estimation time should be outputted.
By default, estimation time is in seconds.
For accepted units, see \link[base:difftime]{base::difftime}.}
}
\value{
A \code{numeric} representing the model estimation time, in seconds.
}
\description{
Get the estimation time of the model, determined by the time taken for the associated \code{\link[=fit]{fit()}} function to finish.
}
\examples{
data(latrendData)
model <- latrend(lcMethodKML("Y", id = "Id", time = "Time"), latrendData)
estimationTime(model)
estimationTime(model, unit = 'mins')
estimationTime(model, unit = 'days')
}
