% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{formula.lcModel}
\alias{formula.lcModel}
\title{Extract the formula of a lcModel}
\usage{
\method{formula}{lcModel}(x, what = "mu", ...)
}
\arguments{
\item{x}{The \code{lcModel} object.}

\item{what}{The distributional parameter}

\item{...}{Additional arguments.}
}
\value{
Returns the associated \code{formula}, or \code{ ~ 0} if not specified.
}
\description{
Extract the formula of a lcModel
}
