% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{getCall.lcModel}
\alias{getCall.lcModel}
\title{Get the model call}
\usage{
\method{getCall}{lcModel}(x, ...)
}
\arguments{
\item{x}{The \code{lcModel} object.}

\item{...}{Not used.}
}
\description{
Get the model call
}
\keyword{internal}
