% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricsExternal.R
\name{getExternalMetricDefinition}
\alias{getExternalMetricDefinition}
\title{Get the external metric definition}
\usage{
getExternalMetricDefinition(name)
}
\arguments{
\item{name}{The name of the metric.}
}
\value{
The metric function, or NULL if not defined.
}
\description{
Get the external metric definition
}
\seealso{
Other metric functions: 
\code{\link{defineExternalMetric}()},
\code{\link{defineInternalMetric}()},
\code{\link{externalMetric}()},
\code{\link{getExternalMetricNames}()},
\code{\link{getInternalMetricDefinition}()},
\code{\link{getInternalMetricNames}()},
\code{\link{metric}()}
}
\concept{metric functions}
