% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{getExternalMetricNames}
\alias{getExternalMetricNames}
\title{Get the names of the available external metrics}
\usage{
getExternalMetricNames()
}
\description{
Get the names of the available external metrics
}
\seealso{
Other metric functions: 
\code{\link{defineExternalMetric}()},
\code{\link{defineInternalMetric}()},
\code{\link{externalMetric,lcModel,lcModel-method}},
\code{\link{getExternalMetricDefinition}()},
\code{\link{getInternalMetricDefinition}()},
\code{\link{getInternalMetricNames}()},
\code{\link{metric,lcModel-method}}
}
\concept{metric functions}
