% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricsInternal.R
\name{getInternalMetricDefinition}
\alias{getInternalMetricDefinition}
\title{Get the internal metric definition}
\usage{
getInternalMetricDefinition(name)
}
\arguments{
\item{name}{The name of the metric.}
}
\value{
The metric function, or NULL if not defined.
}
\description{
Get the internal metric definition
}
\seealso{
Other metric functions: 
\code{\link{defineExternalMetric}()},
\code{\link{defineInternalMetric}()},
\code{\link{externalMetric}()},
\code{\link{getExternalMetricDefinition}()},
\code{\link{getExternalMetricNames}()},
\code{\link{getInternalMetricNames}()},
\code{\link{metric}()}
}
\concept{metric functions}
