% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricsInternal.R
\name{getInternalMetricNames}
\alias{getInternalMetricNames}
\title{Get the names of the available internal metrics}
\usage{
getInternalMetricNames()
}
\description{
Get the names of the available internal metrics
}
\seealso{
Other metric functions: 
\code{\link{defineExternalMetric}()},
\code{\link{defineInternalMetric}()},
\code{\link{externalMetric}()},
\code{\link{getExternalMetricDefinition}()},
\code{\link{getExternalMetricNames}()},
\code{\link{getInternalMetricDefinition}()},
\code{\link{metric}()}
}
\concept{metric functions}
