% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{ids}
\alias{ids}
\title{Get the unique ids included in this model}
\usage{
ids(object)
}
\arguments{
\item{object}{The \code{lcModel} object.}
}
\description{
Get the unique ids included in this model
}
\details{
The order returned by ids(lcModel) determines the id order for any output involving id-specific values, such as in trajectoryAssignments() or postprob()
}
\examples{
model = latrend(lcMethodKML("Y", id = "Id", time = "Time"), latrendData)
ids(model) # S1, S2, ..., S500
}
