% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodAKMedoids.R
\name{interface-akmedoids}
\alias{interface-akmedoids}
\alias{getName,lcMethodAkmedoids-method}
\alias{getShortName,lcMethodAkmedoids-method}
\alias{fit,lcMethodAkmedoids-method}
\title{akmedoids interface}
\usage{
\S4method{getName}{lcMethodAkmedoids}(object)

\S4method{getShortName}{lcMethodAkmedoids}(object)

\S4method{fit}{lcMethodAkmedoids}(method, data, envir, verbose, ...)
}
\arguments{
\item{method}{An object inheriting from \code{lcMethod} with all its arguments having been evaluated and finalized.}

\item{data}{A \code{data.frame} representing the transformed training data.}

\item{verbose}{A \link[R.utils:Verbose]{R.utils::Verbose} object indicating the level of verbosity.}
}
\description{
akmedoids interface
}
\seealso{
\link{lcMethodAkmedoids} \link[akmedoids]{akclustr}
}
\keyword{internal}
