% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodDtwclust.R, R/modelDtwclust.R
\name{interface-dtwclust}
\alias{interface-dtwclust}
\alias{getName,lcMethodDtwclust-method}
\alias{getShortName,lcMethodDtwclust-method}
\alias{preFit,lcMethodDtwclust-method}
\alias{fit,lcMethodDtwclust-method}
\alias{converged,lcModelDtwclust-method}
\alias{postprob,lcModelDtwclust-method}
\alias{predictForCluster,lcModelDtwclust-method}
\title{dtwclust interface}
\usage{
\S4method{getName}{lcMethodDtwclust}(object)

\S4method{getShortName}{lcMethodDtwclust}(object)

\S4method{preFit}{lcMethodDtwclust}(method, data, envir, verbose, ...)

\S4method{fit}{lcMethodDtwclust}(method, data, envir, verbose, ...)

\S4method{converged}{lcModelDtwclust}(object, ...)

\S4method{postprob}{lcModelDtwclust}(object, ...)

\S4method{predictForCluster}{lcModelDtwclust}(object, newdata, cluster, what = "mu", ...)
}
\arguments{
\item{object}{The \code{lcMethod} object.}

\item{method}{The \code{lcMethod} object.}

\item{data}{The data, as a \code{data.frame}, on which the model will be trained.}

\item{envir}{The \code{environment} in which the \code{lcMethod} should be evaluated}

\item{verbose}{A \link[R.utils:Verbose]{R.utils::Verbose} object indicating the level of verbosity.}

\item{...}{Additional arguments.}

\item{newdata}{Optional data frame for which to compute the model predictions. If omitted, the model training data is used.
Cluster trajectory predictions are made when ids are not specified. If the clusters are specified under the Cluster column, output is given only for the specified cluster. Otherwise, a matrix is returned with predictions for all clusters.}

\item{cluster}{The cluster name (as \code{character}) to predict for.}

\item{what}{The distributional parameter to predict. By default, the mean response 'mu' is predicted. The cluster membership predictions can be obtained by specifying what='mb'.}
}
\description{
dtwclust interface
}
\seealso{
\link{lcMethodDtwclust} \link[dtwclust]{dtwclust-package}
}
