% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodFeature.R, R/methodLMKM.R,
%   R/methodGCKM.R, R/methodStratify.R, R/modelLMKM.R
\name{interface-featureBased}
\alias{interface-featureBased}
\alias{getArgumentDefaults,lcMethodFeature-method}
\alias{getName,lcMethodFeature-method}
\alias{getShortName,lcMethodFeature-method}
\alias{prepareData,lcMethodFeature-method}
\alias{fit,lcMethodFeature-method}
\alias{getArgumentDefaults,lcMethodLMKM-method}
\alias{getArgumentExclusions,lcMethodLMKM-method}
\alias{getName,lcMethodLMKM-method}
\alias{getShortName,lcMethodLMKM-method}
\alias{prepareData,lcMethodLMKM-method}
\alias{fit,lcMethodLMKM-method}
\alias{validate,lcMethodLMKM-method}
\alias{getArgumentDefaults,lcMethodGCKM-method}
\alias{getArgumentExclusions,lcMethodGCKM-method}
\alias{getName,lcMethodGCKM-method}
\alias{getShortName,lcMethodGCKM-method}
\alias{prepareData,lcMethodGCKM-method}
\alias{getArgumentDefaults,lcMethodStratify-method}
\alias{getName,lcMethodStratify-method}
\alias{getShortName,lcMethodStratify-method}
\alias{compose,lcMethodStratify-method}
\alias{fit,lcMethodStratify-method}
\alias{coef.lcModelLMKM}
\alias{logLik.lcModelLMKM}
\alias{converged,lcModelLMKM-method}
\alias{postprob,lcModelLMKM-method}
\title{featureBased interface}
\usage{
\S4method{getArgumentDefaults}{lcMethodFeature}(object)

\S4method{getName}{lcMethodFeature}(object)

\S4method{getShortName}{lcMethodFeature}(object)

\S4method{prepareData}{lcMethodFeature}(method, data, verbose, ...)

\S4method{fit}{lcMethodFeature}(method, data, envir, verbose, ...)

\S4method{getArgumentDefaults}{lcMethodLMKM}(object)

\S4method{getArgumentExclusions}{lcMethodLMKM}(object)

\S4method{getName}{lcMethodLMKM}(object)

\S4method{getShortName}{lcMethodLMKM}(object)

\S4method{prepareData}{lcMethodLMKM}(method, data, verbose)

\S4method{fit}{lcMethodLMKM}(method, data, envir, verbose, ...)

\S4method{validate}{lcMethodLMKM}(method, data, envir = NULL, ...)

\S4method{getArgumentDefaults}{lcMethodGCKM}(object)

\S4method{getArgumentExclusions}{lcMethodGCKM}(object)

\S4method{getName}{lcMethodGCKM}(object)

\S4method{getShortName}{lcMethodGCKM}(object)

\S4method{prepareData}{lcMethodGCKM}(method, data, verbose)

\S4method{getArgumentDefaults}{lcMethodStratify}(object)

\S4method{getName}{lcMethodStratify}(object)

\S4method{getShortName}{lcMethodStratify}(object)

\S4method{compose}{lcMethodStratify}(method, envir = NULL, ...)

\S4method{fit}{lcMethodStratify}(method, data, envir, verbose, ...)

\method{coef}{lcModelLMKM}(object, ..., cluster = NULL)

\method{logLik}{lcModelLMKM}(object, ...)

\S4method{converged}{lcModelLMKM}(object, ...)

\S4method{postprob}{lcModelLMKM}(object, ...)
}
\arguments{
\item{object}{The object.}

\item{method}{An object inheriting from \code{lcMethod} with all its arguments having been evaluated and finalized.}

\item{data}{A \code{data.frame} representing the transformed training data.}

\item{verbose}{A \link[R.utils:Verbose]{R.utils::Verbose} object indicating the level of verbosity.}

\item{...}{Not used.}

\item{envir}{The \code{environment} containing variables generated by \code{\link[=prepareData]{prepareData()}} and \code{\link[=preFit]{preFit()}}.}

\item{cluster}{The cluster name.}
}
\description{
featureBased interface
}
\seealso{
\link{lcMethodFeature} \link{lcMethodGCKM} \link{lcMethodLMKM}
}
\keyword{internal}
