% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodFunFEM.R, R/modelFunFEM.R
\name{interface-funFEM}
\alias{interface-funFEM}
\alias{getArgumentDefaults,lcMethodFunFEM-method}
\alias{getArgumentExclusions,lcMethodFunFEM-method}
\alias{getCitation,lcMethodFunFEM-method}
\alias{getName,lcMethodFunFEM-method}
\alias{getShortName,lcMethodFunFEM-method}
\alias{preFit,lcMethodFunFEM-method}
\alias{fit,lcMethodFunFEM-method}
\alias{fitted.lcModelFunFEM}
\alias{predictForCluster,lcModelFunFEM-method}
\alias{postprob,lcModelFunFEM-method}
\alias{coef.lcModelFunFEM}
\alias{logLik.lcModelFunFEM}
\alias{converged,lcModelFunFEM-method}
\title{funFEM interface}
\usage{
\S4method{getArgumentDefaults}{lcMethodFunFEM}(object)

\S4method{getArgumentExclusions}{lcMethodFunFEM}(object)

\S4method{getCitation}{lcMethodFunFEM}(object, ...)

\S4method{getName}{lcMethodFunFEM}(object)

\S4method{getShortName}{lcMethodFunFEM}(object)

\S4method{preFit}{lcMethodFunFEM}(method, data, envir, verbose, ...)

\S4method{fit}{lcMethodFunFEM}(method, data, envir, verbose, ...)

\method{fitted}{lcModelFunFEM}(object, ..., clusters = trajectoryAssignments(object))

\S4method{predictForCluster}{lcModelFunFEM}(
  object,
  newdata,
  cluster,
  what = "mu",
  approxFun = approx,
  ...
)

\S4method{postprob}{lcModelFunFEM}(object, ...)

\method{coef}{lcModelFunFEM}(object, ...)

\method{logLik}{lcModelFunFEM}(object, ...)

\S4method{converged}{lcModelFunFEM}(object, ...)
}
\arguments{
\item{object}{The \code{lcModel} object.}

\item{...}{Not used.}

\item{method}{An object inheriting from \code{lcMethod} with all its arguments having been evaluated and finalized.}

\item{data}{A \code{data.frame} representing the transformed training data.}

\item{envir}{The \code{environment} containing variables generated by \code{\link[=prepareData]{prepareData()}} and \code{\link[=preFit]{preFit()}}.}

\item{verbose}{A \link[R.utils:Verbose]{R.utils::Verbose} object indicating the level of verbosity.}

\item{clusters}{Optional cluster assignments per id. If unspecified, a \code{matrix} is returned containing the cluster-specific predictions per column.}

\item{newdata}{A \code{data.frame} of trajectory data for which to compute trajectory assignments.}

\item{cluster}{The cluster name (as \code{character}) to predict for.}

\item{what}{The distributional parameter to predict. By default, the mean response 'mu' is predicted. The cluster membership predictions can be obtained by specifying \code{what = 'mb'}.}

\item{approxFun}{Function to interpolate between measurement moments, \link[stats]{approx}() by default.}
}
\description{
funFEM interface
}
\seealso{
\link{lcMethodFunFEM} \link[funFEM]{funFEM-package}
}
\keyword{internal}
