% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodKML.R, R/modelKML.R
\docType{class}
\name{interface-kml}
\alias{interface-kml}
\alias{getArgumentDefaults,lcMethodKML-method}
\alias{getArgumentExclusions,lcMethodKML-method}
\alias{getCitation,lcMethodKML-method}
\alias{getName,lcMethodKML-method}
\alias{getShortName,lcMethodKML-method}
\alias{preFit,lcMethodKML-method}
\alias{fit,lcMethodKML-method}
\alias{lcModelKML-class}
\alias{clusterTrajectories,lcModelKML-method}
\alias{converged,lcModelKML-method}
\alias{logLik.lcModelKML}
\alias{postprob,lcModelKML-method}
\alias{predictPostprob,lcModelKML-method}
\title{kml interface}
\usage{
\S4method{getArgumentDefaults}{lcMethodKML}(object)

\S4method{getArgumentExclusions}{lcMethodKML}(object)

\S4method{getCitation}{lcMethodKML}(object, ...)

\S4method{getName}{lcMethodKML}(object)

\S4method{getShortName}{lcMethodKML}(object)

\S4method{preFit}{lcMethodKML}(method, data, envir, verbose, ...)

\S4method{fit}{lcMethodKML}(method, data, envir, verbose, ...)

\S4method{clusterTrajectories}{lcModelKML}(object, at = time(object), ...)

\S4method{converged}{lcModelKML}(object)

\method{logLik}{lcModelKML}(object, ...)

\S4method{postprob}{lcModelKML}(object)

\S4method{predictPostprob}{lcModelKML}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{The model.}

\item{...}{Not used.}

\item{method}{An object inheriting from \code{lcMethod} with all its arguments having been evaluated and finalized.}

\item{data}{A \code{data.frame} representing the transformed training data.}

\item{envir}{The \code{environment} containing additional data variables returned by \code{\link[=prepareData]{prepareData()}}.}

\item{verbose}{A \link[R.utils:Verbose]{R.utils::Verbose} object indicating the level of verbosity.}

\item{at}{A \verb{numeric vector} of the times at which to compute the cluster trajectories.}

\item{newdata}{Optional \code{data.frame} for which to compute the posterior probability. If omitted, the model training data is used.}
}
\description{
kml interface
}
\seealso{
\link{lcMethodKML} \link[kml]{kml}
}
\keyword{internal}
