% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodLcmmGMM.R, R/methodLcmmGBTM.R,
%   R/modelLcmmGMM.R
\name{getName,lcMethodLcmmGMM-method}
\alias{getName,lcMethodLcmmGMM-method}
\alias{getShortName,lcMethodLcmmGMM-method}
\alias{preFit,lcMethodLcmmGMM-method}
\alias{fit,lcMethodLcmmGMM-method}
\alias{interface-lcmm}
\alias{getName,lcMethodLcmmGBTM-method}
\alias{getShortName,lcMethodLcmmGBTM-method}
\alias{preFit,lcMethodLcmmGBTM-method}
\alias{fit,lcMethodLcmmGBTM-method}
\alias{fitted.lcModelLcmmGMM}
\alias{predict.lcModelLcmmGMM}
\alias{model.matrix.lcModelLcmmGMM}
\alias{logLik.lcModelLcmmGMM}
\alias{sigma.lcModelLcmmGMM}
\alias{postprob,lcModelLcmmGMM-method}
\alias{converged,lcModelLcmmGMM-method}
\title{lcmm interface}
\usage{
\S4method{getName}{lcMethodLcmmGMM}(object)

\S4method{getShortName}{lcMethodLcmmGMM}(object)

\S4method{preFit}{lcMethodLcmmGMM}(method, data, envir, verbose, ...)

\S4method{fit}{lcMethodLcmmGMM}(method, data, envir, verbose, ...)

\S4method{getName}{lcMethodLcmmGBTM}(object)

\S4method{getShortName}{lcMethodLcmmGBTM}(object)

\S4method{preFit}{lcMethodLcmmGBTM}(method, data, envir, verbose, ...)

\S4method{fit}{lcMethodLcmmGBTM}(method, data, envir, verbose, ...)

\method{fitted}{lcModelLcmmGMM}(object, ..., clusters = trajectoryAssignments(object))

\method{predict}{lcModelLcmmGMM}(object, ..., newdata = NULL, what = "mu")

\method{model.matrix}{lcModelLcmmGMM}(object, ..., what = "mu")

\method{logLik}{lcModelLcmmGMM}(object, ...)

\method{sigma}{lcModelLcmmGMM}(object, ...)

\S4method{postprob}{lcModelLcmmGMM}(object, ...)

\S4method{converged}{lcModelLcmmGMM}(object, ...)
}
\arguments{
\item{object}{The object to extract the label from.}

\item{method}{The \code{lcMethod} object.}

\item{data}{The data, as a \code{data.frame}, on which the model will be trained.}

\item{envir}{The \code{environment} in which the \code{lcMethod} should be evaluated}

\item{verbose}{A \link[R.utils:Verbose]{R.utils::Verbose} object indicating the level of verbosity.}

\item{...}{Additional arguments.}

\item{clusters}{Optional cluster assignments per id. If unspecified, a \code{matrix} is returned containing the cluster-specific predictions per column.}

\item{newdata}{Optional data frame for which to compute the model predictions. If omitted, the model training data is used.
Cluster trajectory predictions are made when ids are not specified. If the clusters are specified under the Cluster column, output is given only for the specified cluster. Otherwise, a matrix is returned with predictions for all clusters.}

\item{what}{The distributional parameter to predict. By default, the mean response 'mu' is predicted. The cluster membership predictions can be obtained by specifying what='mb'.}
}
\description{
lcmm interface
}
\seealso{
\link{lcMethodLcmmGBTM} \link{lcMethodLcmmGMM} \link[lcmm]{lcmm-package}
}
\keyword{internal}
