% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodLongclust.R, R/modelLongclust.R
\name{interface-longclust}
\alias{interface-longclust}
\alias{getName,lcMethodLongclust-method}
\alias{getShortName,lcMethodLongclust-method}
\alias{fit,lcMethodLongclust-method}
\alias{clusterTrajectories,lcModelLongclust-method}
\alias{postprob,lcModelLongclust-method}
\alias{converged,lcModelLongclust-method}
\alias{logLik.lcModelLongclust}
\alias{BIC.lcModelLongclust}
\title{longclust interface}
\usage{
\S4method{getName}{lcMethodLongclust}(object)

\S4method{getShortName}{lcMethodLongclust}(object)

\S4method{fit}{lcMethodLongclust}(method, data, envir, verbose, ...)

\S4method{clusterTrajectories}{lcModelLongclust}(object, at = time(object), ...)

\S4method{postprob}{lcModelLongclust}(object, ...)

\S4method{converged}{lcModelLongclust}(object, ...)

\method{logLik}{lcModelLongclust}(object, ...)

\method{BIC}{lcModelLongclust}(object, ...)
}
\arguments{
\item{object}{The \code{lcMethod} or \code{lcModel} object.}

\item{method}{An object inheriting from \code{lcMethod} with all its arguments having been evaluated and finalized.}

\item{data}{A \code{data.frame} representing the transformed training data.}

\item{envir}{The \code{environment} containing variables generated by \code{\link[=prepareData]{prepareData()}} and \code{\link[=preFit]{preFit()}}.}

\item{verbose}{A \link[R.utils:Verbose]{R.utils::Verbose} object indicating the level of verbosity.}

\item{...}{Additional arguments.}

\item{at}{An optional vector of the times at which to compute the cluster trajectory predictions.}
}
\description{
longclust interface
}
\seealso{
\link{lcMethodLongclust} \link[longclust]{longclust-package}
}
\keyword{internal}
