% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-method.R, R/meta-fit-converged.R,
%   R/meta-fit-rep.R
\docType{class}
\name{interface-metaMethods}
\alias{interface-metaMethods}
\alias{lcMetaMethod-class}
\alias{compose,lcMetaMethod-method}
\alias{getLcMethod,lcMetaMethod-method}
\alias{getName,lcMetaMethod-method}
\alias{getShortName,lcMetaMethod-method}
\alias{idVariable,lcMetaMethod-method}
\alias{preFit,lcMetaMethod-method}
\alias{prepareData,lcMetaMethod-method}
\alias{fit,lcMetaMethod-method}
\alias{postFit,lcMetaMethod-method}
\alias{responseVariable,lcMetaMethod-method}
\alias{timeVariable,lcMetaMethod-method}
\alias{validate,lcMetaMethod-method}
\alias{update.lcMetaMethod}
\alias{fit,lcFitConverged-method}
\alias{validate,lcFitConverged-method}
\alias{fit,lcFitRep-method}
\alias{validate,lcFitRep-method}
\title{lcMetaMethod abstract class}
\usage{
\S4method{compose}{lcMetaMethod}(method, envir = NULL)

\S4method{getLcMethod}{lcMetaMethod}(object, ...)

\S4method{getName}{lcMetaMethod}(object, ...)

\S4method{getShortName}{lcMetaMethod}(object, ...)

\S4method{idVariable}{lcMetaMethod}(object, ...)

\S4method{preFit}{lcMetaMethod}(method, data, envir, verbose)

\S4method{prepareData}{lcMetaMethod}(method, data, verbose)

\S4method{fit}{lcMetaMethod}(method, data, envir, verbose)

\S4method{postFit}{lcMetaMethod}(method, data, model, envir, verbose)

\S4method{responseVariable}{lcMetaMethod}(object, ...)

\S4method{timeVariable}{lcMetaMethod}(object, ...)

\S4method{validate}{lcMetaMethod}(method, data, envir = NULL, ...)

\method{update}{lcMetaMethod}(object, ...)

\S4method{fit}{lcFitConverged}(method, data, envir, verbose)

\S4method{validate}{lcFitConverged}(method, data, envir = NULL, ...)

\S4method{fit}{lcFitRep}(method, data, envir, verbose)

\S4method{validate}{lcFitRep}(method, data, envir = NULL, ...)
}
\arguments{
\item{method}{The \code{lcMethod} object.}

\item{envir}{The \code{environment} in which the \code{lcMethod} should be evaluated}

\item{object}{The \code{lcModel} object.}

\item{...}{Not used.}

\item{data}{A \code{data.frame} representing the transformed training data.}

\item{verbose}{A \link[R.utils:Verbose]{R.utils::Verbose} object indicating the level of verbosity.}

\item{model}{The \code{lcModel} object returned by \code{\link[=fit]{fit()}}.}
}
\description{
Virtual class for internal use. Do not use.
}
