% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodMixAK_GLMM.R, R/modelMixAK_GLMM.R,
%   R/modelMixAK_GLMMlist.R
\name{interface-mixAK}
\alias{interface-mixAK}
\alias{getName,lcMethodMixAK_GLMM-method}
\alias{getShortName,lcMethodMixAK_GLMM-method}
\alias{responseVariable,lcMethodMixAK_GLMM-method}
\alias{preFit,lcMethodMixAK_GLMM-method}
\alias{fit,lcMethodMixAK_GLMM-method}
\alias{postprob,lcModelMixAK_GLMM-method}
\alias{predictForCluster,lcModelMixAK_GLMM-method}
\alias{coef.lcModelMixAK_GLMM}
\alias{deviance.lcModelMixAK_GLMM}
\alias{postprob,lcModelMixAK_GLMMlist-method}
\alias{predictForCluster,lcModelMixAK_GLMMlist-method}
\title{mixAK interface}
\usage{
\S4method{getName}{lcMethodMixAK_GLMM}(object)

\S4method{getShortName}{lcMethodMixAK_GLMM}(object)

\S4method{responseVariable}{lcMethodMixAK_GLMM}(object)

\S4method{preFit}{lcMethodMixAK_GLMM}(method, data, envir, verbose, ...)

\S4method{fit}{lcMethodMixAK_GLMM}(method, data, envir, verbose, ...)

\S4method{postprob}{lcModelMixAK_GLMM}(object, ...)

\S4method{predictForCluster}{lcModelMixAK_GLMM}(object, newdata, cluster, what = "mu", ...)

\S4method{predictForCluster}{lcModelMixAK_GLMM}(object, newdata, cluster, what = "mu", ...)

\method{coef}{lcModelMixAK_GLMM}(object, ..., stat = "Mean")

\method{deviance}{lcModelMixAK_GLMM}(object, ...)

\S4method{postprob}{lcModelMixAK_GLMMlist}(object, ...)

\S4method{predictForCluster}{lcModelMixAK_GLMMlist}(object, newdata, cluster, what = "mu", ...)
}
\arguments{
\item{object}{The \code{lcMethod} or \code{lcModel} object.}

\item{method}{An object inheriting from \code{lcMethod} with all its arguments having been evaluated and finalized.}

\item{data}{A \code{data.frame} representing the transformed training data.}

\item{verbose}{A \link[R.utils:Verbose]{R.utils::Verbose} object indicating the level of verbosity.}

\item{...}{Additional arguments.}

\item{newdata}{Optional \code{data.frame} for which to compute the model predictions. If omitted, the model training data is used.
Cluster trajectory predictions are made when ids are not specified.}

\item{cluster}{The cluster name (as \code{character}) to predict for.}

\item{what}{The distributional parameter to predict. By default, the mean response 'mu' is predicted. The cluster membership predictions can be obtained by specifying \code{what = 'mb'}.}

\item{stat}{The aggregate statistic to extract. The mean is used by default.}
}
\description{
mixAK interface
}
\seealso{
\link{lcMethodMixAK_GLMM} \link[mixAK]{GLMM_MCMC}
}
\keyword{internal}
