% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R, R/model.R, R/modelCustom.R,
%   R/models.R
\name{latrend-is}
\alias{latrend-is}
\alias{is.lcMethod}
\alias{is.lcModel}
\alias{is.lcModelCustom}
\alias{is.lcModels}
\title{Check if object is of Class}
\usage{
is.lcMethod(x)

is.lcModel(x)

is.lcModelCustom(x)

is.lcModels(x)
}
\arguments{
\item{x}{The object to check the class of.}
}
\description{
Check if object is of Class
}
\keyword{internal}
