% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{isArgDefined}
\alias{isArgDefined}
\title{Check whether the argument of a lcMethod has a defined value.}
\usage{
isArgDefined(object, name, envir = environment(object))
}
\arguments{
\item{object}{The \code{lcMethod} object.}

\item{name}{The name of the argument, as \code{character}.}

\item{envir}{The \code{environment} to evaluate the arguments in. If \code{NULL}, the argument is not evaluated.}
}
\description{
Determines whether the associated argument value is defined. If the argument value is of type \code{language}, the argument is evaluated to see if it can be resolved within its \code{environment}.
}
\keyword{internal}
