% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{latrend-approaches}
\alias{latrend-approaches}
\title{High-level approaches to longitudinal clustering}
\description{
This page provides high-level guidelines on which methods are applicable to your dataset.
Note that this is intended as a quick-start.

Recommended overview and comparison papers:
\itemize{
\item \insertCite{denteuling2021clustering}{latrend}: A tutorial and overview on methods for longitudinal clustering.
\item \insertCite{denteuling2021comparison;textual}{latrend} compared KmL, MixTVEM, GBTM, GMM, and GCKM.
\item \insertCite{twisk2012classifying;textual}{latrend} compared KmL, GCKM, LLCA, GBTM and GMM.
\item \insertCite{verboon2022clustering;textual}{latrend} compared the \emph{kml}, \emph{traj} and \emph{lcmm} packages in R.
\item \insertCite{martin2015growth;textual}{latrend} compared KmL, LCA, and GMM.
}
}
\section{Approaches}{

Disclaimer: The table below has been adapted from a pre-print of \insertCite{denteuling2021clustering}{latrend}.

It is strongly encouraged to \link[=latrend-metrics]{evaluate and compare} several candidate methods in order to identify the most suitable method.
}

\references{
\insertAllCited{}
}
\seealso{
\link{latrend-methods} \link{latrend-estimation} \link{latrend-metrics}
}
