% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/zzz.R
\name{latrend-generics}
\alias{latrend-generics}
\alias{preFit-method}
\title{Method- and model-specific generics defined by the latrend package}
\usage{
compose(method, envir, ...)

fit(method, data, envir, verbose, ...)

postFit(method, data, model, envir, verbose, ...)

prepareData(method, data, verbose, ...)

preFit(method, data, envir, verbose, ...)

validate(method, data, envir, ...)
}
\arguments{
\item{method}{The method.}

\item{envir}{\code{environment}.}

\item{...}{Arguments.}

\item{data}{\code{data.frame}.}

\item{verbose}{\link[R.utils:Verbose]{R.utils::Verbose}.}

\item{object}{The object.}

\item{newdata}{\code{data.frame} of newdata.}

\item{name}{Metric name.}

\item{what}{Parameter.}

\item{cluster}{Cluster name.}
}
\description{
List of S4 generic methods which have no general use other than supporting
functions with signatures of \code{lcMethod} or \code{lcModel}.
}
\keyword{internal}
