% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{latrend-package}
\alias{latrend-package}
\alias{_PACKAGE}
\title{latrend: Framework for clustering longitudinal data}
\description{
A framework for clustering longitudinal datasets with univariate trajectories using commonly applied methods. Facilitates reproducible and transparent cluster analyses with internal and external validation, and simulation studies.
}
\section{Getting started}{

\itemize{
\item See \code{vignette("demo", package="latrend")} for an introduction to conducting a longitudinal cluster analysis on a example case study.
\item See \code{vignette("customModels", package="latrend")} for examples on constructing your own cluster models.
\item See \code{vignette("modelSelection", package="latrend")}
\item See \code{vignette("modelValidation", package="latrend")}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/philips-software/latrend}
  \item Report bugs at \url{https://github.com/philips-software/latrend/issues}
}

}
\author{
\strong{Maintainer}: Niek Den Teuling \email{niek.den.teuling@philips.com} (\href{https://orcid.org/0000-0003-1026-5080}{ORCID})

Other contributors:
\itemize{
  \item Steffen Pauws \email{s.c.pauws@uvt.nl} [contributor]
  \item Edwin van den Heuvel \email{e.r.v.d.heuvel@tue.nl} [contributor]
  \item Copyright © 2020 Koninklijke Philips N.V. [copyright holder]
}

}
