% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latrend.R
\name{latrendRep}
\alias{latrendRep}
\title{Cluster longitudinal data repeatedly}
\usage{
latrendRep(
  method,
  data,
  .rep = 10,
  ...,
  .errorHandling = "stop",
  .seed = NULL,
  .parallel = FALSE,
  envir = NULL,
  verbose = getOption("latrend.verbose")
)
}
\arguments{
\item{method}{An \link{lcMethod} object specifying the longitudinal cluster method to apply, or the name (as \code{character}) of the \code{lcMethod} subclass to instantiate.}

\item{data}{The data of the trajectories to which to estimate the method for.
Any inputs supported by \code{\link[=trajectories]{trajectories()}} can be used, including \code{data.frame} and \code{matrix}.}

\item{.rep}{The number of repeated fits.}

\item{...}{Any other arguments to update the \code{lcMethod} definition with.}

\item{.errorHandling}{Whether to \code{"stop"} on an error, or to \verb{"remove'} evaluations that raised an error.}

\item{.seed}{Set the seed for generating the respective seed for each of the repeated fits.}

\item{.parallel}{Whether to use parallel evaluation. See \link{latrend-parallel}.}

\item{envir}{The \code{environment} in which to evaluate the method arguments via \code{\link[=compose]{compose()}}.
If the \code{data} argument is of type \code{call} then this environment is also used to evaluate the \code{data} argument.}

\item{verbose}{The level of verbosity. Either an object of class \code{Verbose} (see \link[R.utils:Verbose]{R.utils::Verbose} for details),
a \code{logical} indicating whether to show basic computation information,
a \code{numeric} indicating the verbosity level (see \link{Verbose}),
or one of \code{c('info', 'fine', 'finest')}.}
}
\value{
A \code{lcModels} object containing the resulting models.
}
\description{
Performs a repeated fit of the specified latrend model on the given data.
}
\details{
This method is faster than repeatedly calling \link{latrend} as it only prepares the data via \code{prepareData()} once.
}
\examples{
data(latrendData)
method <- lcMethodLMKM(Y ~ Time, id = "Id", time = "Time")
models <- latrendRep(method, data = latrendData, .rep = 5) # 5 repeated runs

models <- latrendRep(method, data = latrendData, .seed = 1, .rep = 3)
}
\seealso{
Other longitudinal cluster fit functions: 
\code{\link{latrend}()},
\code{\link{latrendBatch}()},
\code{\link{latrendBoot}()},
\code{\link{latrendCV}()}
}
\concept{longitudinal cluster fit functions}
