% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelApprox.R
\docType{class}
\name{lcApproxModel-class}
\alias{lcApproxModel-class}
\alias{lcApproxModel}
\alias{fitted.lcApproxModel}
\alias{predictForCluster,lcApproxModel-method}
\title{lcApproxModel class}
\usage{
\method{fitted}{lcApproxModel}(object, ..., clusters = trajectoryAssignments(object))

\S4method{predictForCluster}{lcApproxModel}(
  object,
  newdata,
  cluster,
  what = "mu",
  approxFun = approx,
  ...
)
}
\arguments{
\item{object}{The \code{lcModel} object.}

\item{...}{Additional arguments.}

\item{clusters}{Optional cluster assignments per id. If unspecified, a \code{matrix} is returned containing the cluster-specific predictions per column.}

\item{newdata}{A \code{data.frame} of trajectory data for which to compute trajectory assignments.}

\item{cluster}{The cluster name (as \code{character}) to predict for.}

\item{what}{The distributional parameter to predict. By default, the mean response 'mu' is predicted. The cluster membership predictions can be obtained by specifying \code{what = 'mb'}.}

\item{approxFun}{Function to interpolate between measurement moments, \link[stats]{approx}() by default.}
}
\description{
approx models have defined cluster trajectories at fixed moments in time, which should be interpolated
For a correct implementation, \code{lcApproxModel} requires the extending class to implement \code{clusterTrajectories(at=NULL)}
to return the fixed cluster trajectories
}
