% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-fit-converged.R
\docType{class}
\name{lcFitMethods}
\alias{lcFitMethods}
\alias{lcFitConverged}
\title{Method fit modifiers}
\usage{
lcFitConverged(method, maxRep = Inf)
}
\arguments{
\item{method}{The \code{lcMethod} to use for fitting.}

\item{maxRep}{The maximum number of fit attempts}
}
\description{
A collection of special methods that adapt the fitting procedure of the underlying longitudinal cluster method.
Supported fit methods:
\itemize{
\item \code{lcFitConverged}: Fit a method until a converged result is obtained.
}
}
\examples{
data(latrendData)
method <- lcMethodLMKM(Y ~ Time, id = "Id", time = "Time", nClusters = 2)
metaMethod <- lcFitConverged(method, maxRep = 10)
metaMethod
model <- latrend(metaMethod, latrendData)
}
