% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodMatrix.R
\docType{class}
\name{lcMatrixMethod-class}
\alias{lcMatrixMethod-class}
\alias{getName,lcMatrixMethod-method}
\alias{getShortName,lcMatrixMethod-method}
\alias{prepareData,lcMatrixMethod-method}
\title{lcMatrixMethod}
\usage{
\S4method{getName}{lcMatrixMethod}(object, ...)

\S4method{getShortName}{lcMatrixMethod}(object, ...)

\S4method{prepareData}{lcMatrixMethod}(method, data, verbose, ...)
}
\arguments{
\item{object}{The data object to transform.}

\item{...}{Additional arguments.}

\item{method}{The \code{lcMethod} object.}

\item{data}{The data, as a \code{data.frame}, on which the model will be trained.}

\item{verbose}{A \link[R.utils:Verbose]{R.utils::Verbose} object indicating the level of verbosity.}
}
\description{
lcMatrixMethod
}
\keyword{internal}
