% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-method.R, R/meta-method-converged.R
\name{lcMetaMethod-interface}
\alias{lcMetaMethod-interface}
\alias{getLcMethod,lcMetaMethod-method}
\alias{getName,lcMetaMethod-method}
\alias{getShortName,lcMetaMethod-method}
\alias{preFit,lcMetaMethod-method}
\alias{prepareData,lcMetaMethod-method}
\alias{postFit,lcMetaMethod-method}
\alias{validate,lcMetaMethod-method}
\alias{fit,lcMetaConverged-method}
\alias{validate,lcMetaConverged-method}
\title{lcMetaMethod methods}
\usage{
\S4method{compose}{lcMetaMethod}(method, envir = NULL)

\S4method{getLcMethod}{lcMetaMethod}(object, ...)

\S4method{getName}{lcMetaMethod}(object, ...)

\S4method{getShortName}{lcMetaMethod}(object, ...)

\S4method{preFit}{lcMetaMethod}(method, data, envir, verbose)

\S4method{prepareData}{lcMetaMethod}(method, data, verbose)

\S4method{postFit}{lcMetaMethod}(method, data, model, envir, verbose)

\S4method{validate}{lcMetaMethod}(method, data, envir = NULL, ...)

\S4method{fit}{lcMetaConverged}(method, data, envir, verbose)

\S4method{validate}{lcMetaConverged}(method, data, envir = NULL, ...)
}
\description{
lcMetaMethod methods
}
