% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-method.R, R/meta-method-converged.R
\docType{class}
\name{lcMetaMethods}
\alias{lcMetaMethods}
\alias{lcMetaMethod-class}
\alias{lcMetaConverged-class}
\alias{lcMetaConverged}
\title{Meta methods}
\usage{
lcMetaConverged(method, maxRep = Inf)
}
\arguments{
\item{method}{The \code{lcMethod} to use for fitting.}

\item{maxRep}{The maximum number of fit attempts}
}
\description{
\code{lcMetaMethod} classes are used to specify a repeated or adjusted fitting approach for the given longitudinal cluster method.
Supported meta methods:
\itemize{
\item \code{lcMetaMethodConverged}: Fit a method until a converged result is obtained.
}
}
\examples{
data(latrendData)
method <- lcMethodLMKM(Y ~ Time, id = "Id", time = "Time", nClusters = 2)
metaMethod <- lcMetaConverged(method, maxRep = 10)
metaMethod
model <- latrend(metaMethod, latrendData)
}
