% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{lcMethod}
\alias{lcMethod}
\title{Create a lcMethod object of the specified type and arguments}
\usage{
lcMethod(.class, ..., .defaults = list(), .excludeArgs = c())
}
\arguments{
\item{.class}{The type of \link{lcMethod-class} class}

\item{...}{Any arguments to assign to the method object.}

\item{.defaults}{See \code{defaults} of \link{lcMethod.call}.}

\item{.excludeArgs}{See \code{excludeArgs} of \link{lcMethod.call}.}
}
\description{
Provides a mechanism for creating \code{lcMethod} objects for an arbitrary class.
Note that it is advisable to use the class-specific constructors instead.
}
\seealso{
\link{lcMethod.call}
}
