% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latrend.R
\name{lcModel-data-filters}
\alias{lcModel-data-filters}
\alias{bootSample}
\alias{trainFold}
\alias{testFold}
\title{Data filters for lcModel}
\usage{
bootSample(data, id, seed = NULL)

trainFold(data, fold, id, folds, seed)

testFold(data, fold, id, folds, seed)
}
\arguments{
\item{data}{The \code{data.frame} representing the model dataset.}

\item{id}{The identifier variable name.}

\item{seed}{Optional seed for ensuring reproducibility.}

\item{fold}{The fold to select.}

\item{folds}{Total number of folds to create.}
}
\value{
A subset of \code{data} of type \code{data.frame}.
}
\description{
The data filters are applied by \link{latrend} prior to model estimation. These filters are used in \link{latrendBoot} and \link{latrendCV}.
}
\seealso{
Other validation methods: 
\code{\link{createTestDataFolds}()},
\code{\link{createTestDataFold}()},
\code{\link{createTrainDataFolds}()},
\code{\link{latrendBoot}()},
\code{\link{latrendCV}()}

Other validation methods: 
\code{\link{createTestDataFolds}()},
\code{\link{createTestDataFold}()},
\code{\link{createTrainDataFolds}()},
\code{\link{latrendBoot}()},
\code{\link{latrendCV}()}

Other validation methods: 
\code{\link{createTestDataFolds}()},
\code{\link{createTestDataFold}()},
\code{\link{createTrainDataFolds}()},
\code{\link{latrendBoot}()},
\code{\link{latrendCV}()}
}
\concept{validation methods}
\keyword{internal}
