% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelWeightedPartition.R
\name{lcModelWeightedPartition}
\alias{lcModelWeightedPartition}
\title{Create a lcModel with pre-defined weighted partitioning}
\usage{
lcModelWeightedPartition(
  data,
  response,
  weights,
  center = weighted.meanNA,
  clusterNames = colnames(weights),
  time = getOption("latrend.time"),
  id = getOption("latrend.id"),
  name = "wpart"
)
}
\arguments{
\item{data}{The data on which the cluster result is based, a data.frame.}

\item{response}{The name of the response variable.}

\item{weights}{A \code{numIds} x \code{numClusters} matrix of partition probabilities.}

\item{center}{The \code{function} for computing the longitudinal cluster centers, used for representing the cluster trajectories.}

\item{clusterNames}{The names of the clusters, or a function with input \code{n} outputting a \verb{character vector} of names.}

\item{time}{The name of the time variable.}

\item{id}{The name of the trajectory identification variable.}

\item{name}{The name of the method.}
}
\description{
Create a lcModel with pre-defined weighted partitioning
}
