% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{match.call.all}
\alias{match.call.all}
\title{Argument matching with defaults and parent ellipsis expansion}
\usage{
match.call.all(n = 1L)
}
\arguments{
\item{n}{The number of frames to go back on the calling stack. See \link[base:sys.parent]{base::sys.parent} for more details.}
}
\value{
A \code{call}
}
\description{
Returns a call containing all arguments in specified form, including default arguments.
}
\seealso{
\link[base:match.call]{base::match.call} \link[base:sys.parent]{base::sys.parent}
}
\keyword{internal}
