% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodStratify.R
\name{meanNA}
\alias{meanNA}
\title{Mean ignoring NAs}
\usage{
meanNA(x, ...)
}
\arguments{
\item{x}{An \R object.  Currently there are methods for
    numeric/logical vectors and \link[base:Dates]{date},
    \link[base]{date-time} and \link[base]{time interval} objects.  Complex vectors
    are allowed for \code{trim = 0}, only.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Mean ignoring NAs
}
\keyword{internal}
