% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model.data}
\alias{model.data}
\title{Extract the model training data}
\usage{
model.data(object, ...)
}
\arguments{
\item{object}{The object.}

\item{...}{Additional arguments.}
}
\description{
Extract the model training data
}
\keyword{internal}
