% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{names,lcMethod-method}
\alias{names,lcMethod-method}
\alias{length,lcMethod-method}
\title{lcMethod argument names}
\usage{
\S4method{length}{lcMethod}(x)

\S4method{names}{lcMethod}(x)
}
\arguments{
\item{x}{The \code{lcMethod} object.}
}
\value{
The number of arguments, as \verb{scalar integer}.

A \verb{character vector} of argument names.
}
\description{
Extract the argument names or number of arguments from an \code{lcMethod} object.
}
\examples{
method <- lcMethodLMKM(Y ~ Time)
names(method)
length(method)
}
\seealso{
Other lcMethod functions: 
\code{\link{[[,lcMethod-method}},
\code{\link{as.data.frame.lcMethod}()},
\code{\link{as.data.frame.lcMethods}()},
\code{\link{as.lcMethods}()},
\code{\link{as.list.lcMethod}()},
\code{\link{evaluate.lcMethod}()},
\code{\link{formula.lcMethod}()},
\code{\link{lcMethod-class}},
\code{\link{update.lcMethod}()}
}
\concept{lcMethod functions}
