% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{nobs.lcModel}
\alias{nobs.lcModel}
\title{Number of observations used for the lcModel fit}
\usage{
\method{nobs}{lcModel}(object, ...)
}
\arguments{
\item{object}{The \code{lcModel} object.}

\item{...}{Additional arguments.}
}
\description{
Extracts the number of observations that contributed information towards fitting the cluster trajectories of the respective \code{lcModel} object.
Therefore, only non-missing response observations count towards the number of observations.
}
\examples{
data(latrendData)
method <- lcMethodKML("Y", id = "Id", time = "Time", nClusters = 3)
kml <- latrend(method, latrendData)
nobs(kml)
}
\seealso{
\link{nIds} \link{nClusters}

Other model-specific methods: 
\code{\link{clusterTrajectories}()},
\code{\link{coef.lcModel}()},
\code{\link{converged}()},
\code{\link{deviance.lcModel}()},
\code{\link{df.residual.lcModel}()},
\code{\link{fitted.lcModel}()},
\code{\link{fittedTrajectories}()},
\code{\link{lcModel-class}},
\code{\link{logLik.lcModel}()},
\code{\link{model.frame.lcModel}()},
\code{\link{postprob}()},
\code{\link{predict.lcModel}()},
\code{\link{predictAssignments}()},
\code{\link{predictForCluster}()},
\code{\link{predictPostprob}()},
\code{\link{residuals.lcModel}()},
\code{\link{sigma.lcModel}()},
\code{\link{time.lcModel}()}
}
\concept{model-specific methods}
