% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{OSA.adherence}
\alias{OSA.adherence}
\title{Biweekly Mean Treatment Adherence of OSA Patients over 1 Year}
\format{
An object of class \code{data.frame} with 13000 rows and 5 columns.
}
\usage{
OSA.adherence
}
\description{
Deprecated, renamed to \link{PAP.adh}.
}
\keyword{datasets}
