% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{plot,lcModel,ANY-method}
\alias{plot,lcModel,ANY-method}
\title{Plot a lcModel}
\usage{
\S4method{plot}{lcModel,ANY}(x, y, ...)
}
\arguments{
\item{x}{The \code{lcModel} object.}

\item{y}{Not used.}

\item{...}{Arguments passed to \link{plotClusterTrajectories}.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot a lcModel. By default, the cluster trajectories of the model solution are plotted.
}
