% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{plot,lcModels,ANY-method}
\alias{plot,lcModels,ANY-method}
\title{Grid plot for a list of models}
\usage{
\S4method{plot}{lcModels,ANY}(x, y, ..., subset, gridArgs = list())
}
\arguments{
\item{x}{The \code{lcModels} object.}

\item{y}{Not used.}

\item{...}{Additional parameters passed to the \code{plot()} call for each \code{lcModel} object.}

\item{subset}{Logical expression based on the \code{lcModel} method arguments, indicating
which \code{lcModel} objects to keep.}

\item{gridArgs}{Named list of parameters passed to \link[gridExtra:arrangeGrob]{gridExtra::arrangeGrob}.}
}
\description{
Grid plot for a list of models
}
