% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/trajectories.R
\name{plotTrajectories}
\alias{plotTrajectories}
\alias{plotTrajectories,data.frame-method}
\alias{plotTrajectories,ANY-method}
\title{Plot the data trajectories}
\usage{
plotTrajectories(object, ...)

\S4method{plotTrajectories}{data.frame}(
  object,
  response,
  time = getOption("latrend.time"),
  id = getOption("latrend.id"),
  cluster = NULL,
  facet = TRUE,
  ...
)

\S4method{plotTrajectories}{ANY}(object, ...)
}
\arguments{
\item{object}{The data or model or extract the trajectories from.}

\item{response}{Response variable \code{character} name or a \code{call}.}

\item{time}{The time variable name.}

\item{id}{The identifier variable name.}

\item{cluster}{Cluster variable name. If unspecified, trajectories are not grouped. Alternatively, cluster is a vector indicating cluster membership per id.}

\item{facet}{Whether to facet by cluster.}
}
\description{
Plots the output of \link{trajectories} for the given object.
}
\examples{
data(latrendData)
plotTrajectories(latrendData, response = "Y", id = "Id", time = "Time")

plotTrajectories(latrendData, response = quote(exp(Y)), id = "Id", time = "Time")
}
