% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{postprob}
\alias{postprob}
\alias{postprob,lcModel-method}
\title{Posterior probability per fitted trajectory}
\usage{
\S4method{postprob}{lcModel}(object, ...)
}
\arguments{
\item{object}{The \code{lcModel}.}

\item{...}{Additional arguments.}
}
\value{
A I-by-K \code{matrix} with \code{I = nIds(object)} and \code{K = nClusters(object)}.
}
\description{
Get the posterior probability matrix with element \eqn{(i,j)} indicating the probability of trajectory \eqn{i} belonging to cluster \eqn{j}.
}
\details{
This method should be extended by \code{lcModel} implementations. The default implementation returns uniform probabilities for all observations.
}
\examples{
library(lcmm)
data(latrendData)
model <- latrend(lcMethodLcmmGMM(fixed = Y ~ Time, mixture = ~ Time,
   id = "Id", time = "Time"), data = latrendData)
postprob(model)
}
\seealso{
\link{trajectoryAssignments} \link{predictPostprob} \link{predictAssignments}

Other model-specific methods: 
\code{\link{clusterTrajectories}()},
\code{\link{coef.lcModel}()},
\code{\link{converged}()},
\code{\link{deviance.lcModel}()},
\code{\link{df.residual.lcModel}()},
\code{\link{fitted.lcModel}()},
\code{\link{lcModel-class}},
\code{\link{logLik.lcModel}()},
\code{\link{model.frame.lcModel}()},
\code{\link{nobs.lcModel}()},
\code{\link{predict.lcModel}()},
\code{\link{predictAssignments}()},
\code{\link{predictForCluster}()},
\code{\link{predictPostprob}()},
\code{\link{residuals.lcModel}()},
\code{\link{sigma.lcModel}()},
\code{\link{time.lcModel}()},
\code{\link{trajectories}()}
}
\concept{model-specific methods}
