% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodStratify.R
\name{postprobFromAssignments}
\alias{postprobFromAssignments}
\title{Create a posterior probability matrix from a vector of cluster assignments.}
\usage{
postprobFromAssignments(assignments, k)
}
\arguments{
\item{assignments}{Integer vector indicating cluster assignment per trajectory}

\item{k}{The number of clusters.}
}
\description{
For each trajectory, the probability of the assigned cluster is 1.
}
