% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{postProbFromObs}
\alias{postProbFromObs}
\title{Compute the id-specific postprob matrix from a given observation-level postprob matrix}
\usage{
postProbFromObs(mat, rowIds)
}
\arguments{
\item{mat}{A posterior probability \code{matrix} at the observation level.}

\item{rowIds}{The row trajectory identifier \code{integer} vector.}
}
\description{
Compute the id-specific postprob matrix from a given observation-level postprob matrix
}
\keyword{internal}
