% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{summary.lcModel}
\alias{summary.lcModel}
\title{Summarize a lcModel}
\usage{
\method{summary}{lcModel}(object, ...)
}
\arguments{
\item{object}{The \code{lcModel} object.}

\item{...}{Additional arguments.}
}
\description{
Extracts all relevant information from the underlying model into a list
}
