% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{testLcMethod}
\alias{testLcMethod}
\title{Test a lcMethod implementation}
\usage{
testLcMethod(class, instantiator = NULL, args = list())
}
\arguments{
\item{class}{The name of the \code{class} inheriting from \code{lcMethod} to test.}

\item{instantiator}{A \code{function} with signature \verb{(id, time, response)}}

\item{args}{Other arguments passed to the instantiator function.
returning an object inheriting from the \code{lcMethod} specified by the \code{class} argument.}
}
\description{
Test a \code{lcMethod} subclass implementation including the outputted \code{lcModel} class.
}
\note{
This is an experimental function that is subject to large changes in the future.
}
\examples{
testLcMethod("lcMethodRandom")
}
