% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{time.lcModel}
\alias{time.lcModel}
\title{Sampling times of a lcModel}
\usage{
\method{time}{lcModel}(x, ...)
}
\arguments{
\item{x}{The \code{lcModel} object.}

\item{...}{Not used.}
}
\value{
The unique times at which observations occur.
}
\description{
Sampling times of a lcModel
}
\seealso{
Other model-specific methods: 
\code{\link{clusterTrajectories}()},
\code{\link{coef.lcModel}()},
\code{\link{converged}()},
\code{\link{deviance.lcModel}()},
\code{\link{df.residual.lcModel}()},
\code{\link{fitted.lcModel}()},
\code{\link{lcModel-class}},
\code{\link{logLik.lcModel}()},
\code{\link{model.frame.lcModel}()},
\code{\link{nobs.lcModel}()},
\code{\link{postprob}()},
\code{\link{predict.lcModel}()},
\code{\link{predictAssignments}()},
\code{\link{predictForCluster}()},
\code{\link{predictPostprob}()},
\code{\link{residuals.lcModel}()},
\code{\link{sigma.lcModel}()},
\code{\link{trajectories}()}
}
\concept{model-specific methods}
