% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R, R/meta-method.R, R/model.R
\name{timeVariable}
\alias{timeVariable}
\alias{timeVariable,lcMethod-method}
\alias{timeVariable,lcMetaMethod-method}
\alias{timeVariable,lcModel-method}
\title{Extract the time variable}
\usage{
\S4method{timeVariable}{lcMethod}(object, ...)

\S4method{timeVariable}{lcMetaMethod}(object, ...)

\S4method{timeVariable}{lcModel}(object)
}
\arguments{
\item{object}{The object to extract the variable from.}

\item{...}{Additional arguments.}
}
\value{
The time variable name, as \code{character}.
}
\description{
Extracts the time variable (i.e., column name) from the given \code{object}.
}
\examples{
method <- lcMethodLMKM(Y ~ Time, id = "Id", time = "Time")
timeVariable(method) # "Time"
data(latrendData)
method <- lcMethodRandom("Y", id = "Id", time = "Time")
model <- latrend(method, latrendData)
timeVariable(model) # "Time"
}
\seealso{
Other lcModel variables: 
\code{\link{idVariable}()},
\code{\link{responseVariable}()}
}
\concept{lcModel variables}
