% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-transform.R
\name{transformFitted}
\alias{transformFitted}
\alias{transformFitted,NULL,lcModel-method}
\alias{transformFitted,matrix,lcModel-method}
\alias{transformFitted,list,lcModel-method}
\alias{transformFitted,data.frame,lcModel-method}
\title{Helper function for ensuring the right fitted() output}
\usage{
transformFitted(pred, model, clusters)

\S4method{transformFitted}{`NULL`,lcModel}(pred, model, clusters)

\S4method{transformFitted}{matrix,lcModel}(pred, model, clusters)

\S4method{transformFitted}{list,lcModel}(pred, model, clusters)

\S4method{transformFitted}{data.frame,lcModel}(pred, model, clusters)
}
\arguments{
\item{pred}{Prediction object}

\item{model}{The model from which the prediction is made.}

\item{clusters}{Optional argument for specifying the trajectory cluster assignments.}
}
\value{
A \code{vector} if the \code{clusters} argument is specified, else a \code{matrix} with the fitted values per cluster per column.
}
\description{
This function is also responsible for checking whether the input data is valid, such that the fitting process can fail early.
}
