% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-transform.R
\name{transformPredict}
\alias{transformPredict}
\alias{transformPredict,NULL,lcModel-method}
\alias{transformPredict,vector,lcModel-method}
\alias{transformPredict,matrix,lcModel-method}
\alias{transformPredict,data.frame,lcModel-method}
\title{Helper function that matches the output to the specified newdata}
\usage{
transformPredict(pred, model, newdata)

\S4method{transformPredict}{`NULL`,lcModel}(pred, model, newdata)

\S4method{transformPredict}{vector,lcModel}(pred, model, newdata)

\S4method{transformPredict}{matrix,lcModel}(pred, model, newdata)

\S4method{transformPredict}{data.frame,lcModel}(pred, model, newdata)
}
\arguments{
\item{pred}{The prediction object}

\item{model}{The model for which the prediction is made.}

\item{newdata}{A \code{data.frame} containing the input data to predict for.}
}
\value{
A data.frame with the predictions, or a list of cluster-specific prediction frames
}
\description{
If Cluster is not provided, the prediction is outputted in long format per cluster,
resulting in a longer data.frame than the newdata input
}
