% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{update.lcModel}
\alias{update.lcModel}
\title{Update a lcModel}
\usage{
\method{update}{lcModel}(object, ...)
}
\arguments{
\item{object}{The \code{lcModel} object.}

\item{...}{
  Arguments passed on to \code{\link[=latrend]{latrend}}
  \describe{
    \item{\code{method}}{The \code{lcMethod} object specifying the longitudinal cluster method to apply.}
    \item{\code{data}}{The \code{data.frame} or \code{matrix} to which to apply the method.}
    \item{\code{envir}}{The \code{environment} in which to evaluate the method arguments. Note that this only applies to \code{data} when \code{data} is a \code{call}.}
    \item{\code{verbose}}{The level of verbosity. Either an object of class \code{Verbose} (see \link[R.utils:Verbose]{R.utils::Verbose} for details),
a \code{logical} indicating whether to show basic computation information,
a \code{numeric} indicating the verbosity level (see \link{Verbose}),
or one of \code{c('info', 'fine', 'finest')}.}
  }}
}
\value{
The refitted \code{lcModel} object, of the same type as the \code{object} argument.
}
\description{
Fit a new model with modified arguments from the current model.
}
\examples{
library(kml)
data(latrendData)
m <- lcMethodKML("Y", id = "Id", time = "Time", nClusters = 3)
model <- latrend(method = m, data = latrendData)
# fit for a different number of clusters
update(model, nClusters = 2)
}
\seealso{
\link{latrend} \link{getCall}
}
