% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{validate}
\alias{validate}
\alias{validate,lcMethod-method}
\title{Validate the argument values of a lcMethod object}
\usage{
\S4method{validate}{lcMethod}(method, data, envir = NULL, ...)
}
\arguments{
\item{method}{An object inheriting from \code{lcMethod} with all its arguments having been evaluated and finalized.}

\item{data}{A \code{data.frame} representing the transformed training data.}

\item{envir}{The \code{environment} in which the \code{lcMethod} should be evaluated}

\item{...}{Not used.}
}
\value{
Either \code{TRUE} if all validation checks passed,
or a \code{character} containing a description of the failed validation checks.
}
\description{
Validate the argument values of a lcMethod object
}
\section{Implementation}{

An example implementation checking for the existence of specific arguments and type:
\preformatted{
library(assertthat)
setMethod("validate", "lcMethodExample", function(method, data, envir = NULL, ...) {
  validate_that(
    hasName(method, "myArgument"),
    hasName(method, "anotherArgument"),
    is.numeric(method$myArgument)
  )
})
}
}

\section{Fitting procedure}{

The fitting procedure for \code{lcMethod} objects is handled through a series of steps implemented by the \code{lcMethod} object.
The fitting procedure is handled by \code{\link[=latrend]{latrend()}}, in executed in the following order:
\enumerate{
\item \code{\link[=compose]{compose()}}: Evaluate and finalize the method argument values.
\item \code{\link[=validate]{validate()}}: Check the validity of the method argument values.
\item \code{\link[=prepareData]{prepareData()}}: Process the training data for fitting.
\item \code{\link[=preFit]{preFit()}}: Prepare environment for estimation, independent of training data.
\item \code{\link[=fit]{fit()}}: Estimate the specified method on the training data, outputting an object inheriting from \code{lcModel}.
\item \code{\link[=postFit]{postFit()}}: Post-process the outputted \code{lcModel} object.
}

The result of the fitting procedure is an \linkS4class{lcModel} object that inherits from the \code{lcModel} class.
}

\seealso{
\link[assertthat:validate_that]{assertthat::validate_that}
}
