% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodStratify.R
\name{weighted.meanNA}
\alias{weighted.meanNA}
\title{Weighted arithmetic mean ignoring NAs}
\usage{
weighted.meanNA(x, w, ...)
}
\arguments{
\item{x}{an object containing the values whose weighted mean is to be
    computed.}

\item{w}{a numerical vector of weights the same length as \code{x} giving
    the weights to use for elements of \code{x}.}

\item{...}{arguments to be passed to or from methods.}
}
\description{
Weighted arithmetic mean ignoring NAs
}
\keyword{internal}
