% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{which.weight}
\alias{which.weight}
\title{Sample an index of a vector weighted by the elements}
\usage{
which.weight(x)
}
\arguments{
\item{x}{A positive \verb{numeric vector}.}
}
\value{
An \code{integer} giving the index of the sampled element.
}
\description{
Returns a random index, weighted by the element magnitudes. This function is intended to be used as an optional strategy for \link{trajectoryAssignments,lcModel-method}, resulting in randomly sampled cluster membership.
}
\examples{
x = c(.01, .69, .3)
which.weight(x) #1, 2, or 3
}
