% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{as.lcModels}
\alias{as.lcModels}
\title{Convert a list of lcModels to a lcModels list}
\usage{
as.lcModels(x)
}
\arguments{
\item{x}{A \code{list} of \code{lcModel} objects, an \code{lcModels} object, or \code{NULL}.}
}
\value{
A \code{lcModels} object.
}
\description{
Convert a list of lcModels to a lcModels list
}
\seealso{
lcModels

Other lcModel list functions: 
\code{\link{lcModels}},
\code{\link{print.lcModels}()},
\code{\link{subset.lcModels}()}
}
\concept{lcModel list functions}
