% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{clusterNames<-}
\alias{clusterNames<-}
\title{Update the cluster names}
\usage{
clusterNames(object) <- value
}
\arguments{
\item{object}{The \code{lcModel} object to update.}

\item{value}{The \code{character} with the new names.}
}
\value{
The updated \code{lcModel} object.
}
\description{
Update the cluster names
}
\examples{
data(latrendData)
method <- lcMethodLMKM(Y ~ Time, id = "Id", time = "Time")
model <- latrend(method, latrendData, nClusters = 2)
clusterNames(model) <- c("Group 1", "Group 2")
}
