% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{clusterProportions}
\alias{clusterProportions}
\alias{clusterProportions,lcModel-method}
\title{Proportional size of each cluster}
\usage{
\S4method{clusterProportions}{lcModel}(object, ...)
}
\arguments{
\item{object}{The \code{lcModel} to obtain the proportions from.}

\item{...}{Not used.}
}
\description{
Proportional size of each cluster
}
\examples{
data(latrendData)
model <- latrend(lcMethodKML("Y", id = "Id", time = "Time"), latrendData)
clusterProportions(model)
}
