% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{clusterSizes}
\alias{clusterSizes}
\title{Number of strata per cluster}
\usage{
clusterSizes(object)
}
\arguments{
\item{object}{The \code{lcModel} object.}
}
\description{
Number of strata per cluster
}
\examples{
model <- latrend(lcMethodKML("Y", id = "Id", time = "Time"), latrendData)
clusterSizes(model)
}
