% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{coef.lcModel}
\alias{coef.lcModel}
\title{Coefficients of a lcModel}
\usage{
\method{coef}{lcModel}(object, ...)
}
\arguments{
\item{object}{The \code{lcModel} object.}

\item{...}{Additional arguments.}
}
\value{
A \verb{named numeric vector} with all coefficients, or a \code{matrix} with each column containing the cluster-specific coefficients.
}
\description{
Coefficients of a lcModel
}
\seealso{
Other model-specific methods: 
\code{\link{clusterTrajectories}()},
\code{\link{converged}()},
\code{\link{deviance.lcModel}()},
\code{\link{df.residual.lcModel}()},
\code{\link{fitted.lcModel}()},
\code{\link{lcModel-class}},
\code{\link{logLik.lcModel}()},
\code{\link{model.frame.lcModel}()},
\code{\link{nobs.lcModel}()},
\code{\link{postprob}()},
\code{\link{predict.lcModel}()},
\code{\link{predictAssignments}()},
\code{\link{predictForCluster}()},
\code{\link{predictPostprob}()},
\code{\link{residuals.lcModel}()},
\code{\link{sigma.lcModel}()},
\code{\link{time.lcModel}()},
\code{\link{trajectories}()}
}
\concept{model-specific methods}
