% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{confusionMatrix}
\alias{confusionMatrix}
\alias{confusionMatrix.lcModel}
\title{Posterior confusion matrix}
\usage{
confusionMatrix.lcModel(data, ...)
}
\arguments{
\item{data}{The \code{lcModel} object.}

\item{...}{Not used.}
}
\description{
Posterior confusion matrix
}
\examples{
data(latrendData)
model = latrend(method=lcMethodLcmmGMM(Y ~ Time + (1 | Id), id = "Id", time = "Time"),
  data=latrendData)
confusionMatrix(model)
}
\seealso{
\link[caret:confusionMatrix]{caret::confusionMatrix}
}
