% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{deviance.lcModel}
\alias{deviance.lcModel}
\title{lcModel deviance}
\usage{
\method{deviance}{lcModel}(object, ...)
}
\arguments{
\item{object}{The \code{lcModel} object.}

\item{...}{Additional arguments.}
}
\value{
A \code{numeric} with the deviance value. If unavailable, \code{NA} is returned.
}
\description{
Get the deviance of the fitted \code{lcModel} object.
}
\details{
The default implementation checks for the existence of the \code{deviance()} function for the internal model, and returns the output, if available.
}
\seealso{
\link[stats:deviance]{stats::deviance} \link{metric}

Other model-specific methods: 
\code{\link{clusterTrajectories}()},
\code{\link{coef.lcModel}()},
\code{\link{converged}()},
\code{\link{df.residual.lcModel}()},
\code{\link{fitted.lcModel}()},
\code{\link{lcModel-class}},
\code{\link{logLik.lcModel}()},
\code{\link{model.frame.lcModel}()},
\code{\link{nobs.lcModel}()},
\code{\link{postprob}()},
\code{\link{predict.lcModel}()},
\code{\link{predictAssignments}()},
\code{\link{predictForCluster}()},
\code{\link{predictPostprob}()},
\code{\link{residuals.lcModel}()},
\code{\link{sigma.lcModel}()},
\code{\link{time.lcModel}()},
\code{\link{trajectories}()}
}
\concept{model-specific methods}
