% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{df.residual.lcModel}
\alias{df.residual.lcModel}
\title{Extract the residual degrees of freedom from a lcModel}
\usage{
\method{df.residual}{lcModel}(object, ...)
}
\arguments{
\item{object}{The \code{lcModel} object.}

\item{...}{Additional arguments.}
}
\value{
A \code{numeric} with the residual degrees of freedom. If unavailable, \code{NA} is returned.
}
\description{
Extract the residual degrees of freedom from a lcModel
}
\seealso{
\link[stats:df.residual]{stats::df.residual} \link{nobs} \link{residuals}

Other model-specific methods: 
\code{\link{clusterTrajectories}()},
\code{\link{coef.lcModel}()},
\code{\link{converged}()},
\code{\link{deviance.lcModel}()},
\code{\link{fitted.lcModel}()},
\code{\link{lcModel-class}},
\code{\link{logLik.lcModel}()},
\code{\link{model.frame.lcModel}()},
\code{\link{nobs.lcModel}()},
\code{\link{postprob}()},
\code{\link{predict.lcModel}()},
\code{\link{predictAssignments}()},
\code{\link{predictForCluster}()},
\code{\link{predictPostprob}()},
\code{\link{residuals.lcModel}()},
\code{\link{sigma.lcModel}()},
\code{\link{time.lcModel}()},
\code{\link{trajectories}()}
}
\concept{model-specific methods}
