% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{estimationTime}
\alias{estimationTime}
\title{Get the model estimation time}
\usage{
estimationTime(object)
}
\arguments{
\item{object}{The \code{lcModel} object.}
}
\value{
The model estimation time in seconds.
}
\description{
Get the model estimation time
}
