% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{estimationTime}
\alias{estimationTime}
\title{Get the model estimation time}
\usage{
estimationTime(object)
}
\arguments{
\item{object}{The \code{lcModel} object.}
}
\value{
A \code{numeric} representing the model estimation time, in seconds.
}
\description{
Get the estimation time of the model, determined by the time taken for the associated \code{\link[=fit]{fit()}} function to finish.
}
