% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{fit}
\alias{fit}
\alias{fit,lcMethod-method}
\title{Fit an lcMethod object to the processed data}
\usage{
\S4method{fit}{lcMethod}(method, data, envir, verbose)
}
\arguments{
\item{method}{An object inheriting from \code{lcMethod} with all its arguments having been evaluated and finalized.}

\item{data}{A \code{data.frame} representing the transformed training data.}

\item{verbose}{A \link[R.utils:Verbose]{R.utils::Verbose} object indicating the level of verbosity.}
}
\value{
The fitted object inheriting from \code{lcModel}.
}
\description{
Estimates the model as determined by the evaluated method specification, processed training data, and prepared environment.
}
\section{Implementation}{

This method should be implemented for all \code{lcMethod} subclasses.

\preformatted{
setMethod("fit", "lcMethodExample", function(method, data, envir, verbose) {
  # estimate the model or cluster parameters
  coefs <- FIT_CODE

  # create the lcModel object
  new("lcModelExample",
    method = method,
    data = data,
    model = coefs,
    clusterNames = make.clusterNames(method$nClusters)
  )
}
}
}

\section{Fitting procedure}{

The fitting procedure for \code{lcMethod} objects is handled through a series of steps implemented by the \code{lcMethod} object.
The fitting procedure is handled by \code{\link[=latrend]{latrend()}}, in executed in the following order:
\enumerate{
\item \code{\link[=compose]{compose()}}: Evaluate and finalize the method argument values.
\item \code{\link[=validate]{validate()}}: Check the validity of the method argument values.
\item \code{\link[=prepareData]{prepareData()}}: Process the training data for fitting.
\item \code{\link[=preFit]{preFit()}}: Prepare environment for estimation, independent of training data.
\item \code{\link[=fit]{fit()}}: Estimate the specified method on the training data, outputting an object inheriting from \code{lcModel}.
\item \code{\link[=postFit]{postFit()}}: Post-process the outputted \code{lcModel} object.
}

The result of the fitting procedure is an \linkS4class{lcModel} object that inherits from the \code{lcModel} class.
}

