% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{getLcMethod}
\alias{getLcMethod}
\title{Get the method specification of a lcModel}
\usage{
getLcMethod(object)
}
\arguments{
\item{object}{The \code{lcModel} object.}
}
\description{
Get the method specification of a lcModel
}
\examples{
model = latrend(method=lcMethodKML("Y", id = "Id", time = "Time"), data=latrendData)
getLcMethod(model)
}
