% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{getLcMethod}
\alias{getLcMethod}
\title{Get the method specification of a lcModel}
\usage{
getLcMethod(object)
}
\arguments{
\item{object}{The \code{lcModel} object.}
}
\value{
An \code{lcMethod} object.
}
\description{
Get the \code{lcMethod} specification object that was used for fitting the given \code{lcModel} object.
}
\examples{
model = latrend(method=lcMethodKML("Y", id = "Id", time = "Time"), data=latrendData)
getLcMethod(model)
}
\seealso{
\link{getCall.lcModel}
}
