% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{initialize,lcMethod-method}
\alias{initialize,lcMethod-method}
\title{lcMethod initialization}
\usage{
\S4method{initialize}{lcMethod}(.Object, ...)
}
\arguments{
\item{.Object}{The newly allocated \code{lcMethod} object.}

\item{...}{Other method arguments.}
}
\description{
Initialization of \code{lcMethod} objects, converting arbitrary arguments to arguments as part of an \code{lcMethod} object.
}
\examples{
new("lcMethodKML", response = "Value")
}
