% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodAKMedoids.R
\name{interface-akmedoids}
\alias{interface-akmedoids}
\alias{getArgumentDefaults,lcMethodAkmedoids-method}
\alias{getArgumentExclusions,lcMethodAkmedoids-method}
\alias{getCitation,lcMethodAkmedoids-method}
\alias{getName,lcMethodAkmedoids-method}
\alias{getShortName,lcMethodAkmedoids-method}
\alias{fit,lcMethodAkmedoids-method}
\title{akmedoids interface}
\usage{
\S4method{getArgumentDefaults}{lcMethodAkmedoids}(object)

\S4method{getArgumentExclusions}{lcMethodAkmedoids}(object)

\S4method{getCitation}{lcMethodAkmedoids}(object, ...)

\S4method{getName}{lcMethodAkmedoids}(object)

\S4method{getShortName}{lcMethodAkmedoids}(object)

\S4method{fit}{lcMethodAkmedoids}(method, data, envir, verbose, ...)
}
\arguments{
\item{...}{Not used.}

\item{method}{An object inheriting from \code{lcMethod} with all its arguments having been evaluated and finalized.}

\item{data}{A \code{data.frame} representing the transformed training data.}

\item{envir}{The \code{environment} containing variables generated by \code{\link[=prepareData]{prepareData()}} and \code{\link[=preFit]{preFit()}}.}

\item{verbose}{A \link[R.utils:Verbose]{R.utils::Verbose} object indicating the level of verbosity.}
}
\description{
akmedoids interface
}
\seealso{
\link{lcMethodAkmedoids} \code{akmedoids::akclustr}
}
\keyword{internal}
