% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodAKMedoids.R
\name{interface-akmedoids}
\alias{interface-akmedoids}
\alias{getName,lcMethodAkmedoids-method}
\alias{getShortName,lcMethodAkmedoids-method}
\alias{fit,lcMethodAkmedoids-method}
\title{akmedoids interface}
\usage{
\S4method{getName}{lcMethodAkmedoids}(object)

\S4method{getShortName}{lcMethodAkmedoids}(object)

\S4method{fit}{lcMethodAkmedoids}(method, data, envir, verbose, ...)
}
\arguments{
\item{object}{The object to extract the label from.}

\item{method}{The \code{lcMethod} object.}

\item{data}{The data, as a \code{data.frame}, on which the model will be trained.}

\item{envir}{The \code{environment} in which the \code{lcMethod} should be evaluated}

\item{verbose}{A \link[R.utils:Verbose]{R.utils::Verbose} object indicating the level of verbosity.}

\item{...}{Additional arguments.}
}
\description{
akmedoids interface
}
\seealso{
\link{lcMethodAkmedoids} \link[akmedoids]{akclustr}
}
\keyword{internal}
