% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodCrimCV.R, R/modelCrimCV.R
\name{interface-crimCV}
\alias{interface-crimCV}
\alias{getName,lcMethodCrimCV-method}
\alias{getShortName,lcMethodCrimCV-method}
\alias{prepareData,lcMethodCrimCV-method}
\alias{fit,lcMethodCrimCV-method}
\alias{predictForCluster,lcModelCrimCV-method}
\alias{postprob,lcModelCrimCV-method}
\alias{logLik.lcModelCrimCV}
\alias{coef.lcModelCrimCV}
\alias{converged,lcModelCrimCV-method}
\title{crimCV interface}
\usage{
\S4method{getName}{lcMethodCrimCV}(object)

\S4method{getShortName}{lcMethodCrimCV}(object)

\S4method{prepareData}{lcMethodCrimCV}(method, data, verbose, ...)

\S4method{fit}{lcMethodCrimCV}(method, data, envir, verbose, ...)

\S4method{predictForCluster}{lcModelCrimCV}(object, newdata, cluster, what = "mu", ...)

\S4method{postprob}{lcModelCrimCV}(object)

\method{logLik}{lcModelCrimCV}(object, ...)

\method{coef}{lcModelCrimCV}(object, ...)

\S4method{converged}{lcModelCrimCV}(object)
}
\arguments{
\item{object}{The \code{lcMethod} or \code{lcModel} object.}

\item{method}{An object inheriting from \code{lcMethod} with all its arguments having been evaluated and finalized.}

\item{data}{A \code{data.frame} representing the transformed training data.}

\item{verbose}{A \link[R.utils:Verbose]{R.utils::Verbose} object indicating the level of verbosity.}

\item{...}{Additional arguments.}

\item{newdata}{Optional \code{data.frame} for which to compute the model predictions. If omitted, the model training data is used.
Cluster trajectory predictions are made when ids are not specified.}

\item{cluster}{The cluster name (as \code{character}) to predict for.}

\item{what}{The distributional parameter to predict. By default, the mean response 'mu' is predicted. The cluster membership predictions can be obtained by specifying \code{what = 'mb'}.}
}
\description{
crimCV interface
}
\seealso{
\link{lcMethodCrimCV} \link[crimCV]{crimCV}
}
\keyword{internal}
