% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodTwoStep.R, R/methodGCKM.R,
%   R/methodLMKM.R, R/methodStratify.R, R/modelLMKM.R, R/modelTwoStep.R
\name{interface-featureBased}
\alias{interface-featureBased}
\alias{getName,lcMethodTwoStep-method}
\alias{getShortName,lcMethodTwoStep-method}
\alias{prepareData,lcMethodTwoStep-method}
\alias{fit,lcMethodTwoStep-method}
\alias{getName,lcMethodGCKM-method}
\alias{getShortName,lcMethodGCKM-method}
\alias{compose,lcMethodGCKM-method}
\alias{preFit,lcMethodGCKM-method}
\alias{fit,lcMethodGCKM-method}
\alias{getName,lcMethodLMKM-method}
\alias{getShortName,lcMethodLMKM-method}
\alias{prepareData,lcMethodLMKM-method}
\alias{fit,lcMethodLMKM-method}
\alias{getName,lcMethodStratify-method}
\alias{getShortName,lcMethodStratify-method}
\alias{compose,lcMethodStratify-method}
\alias{fit,lcMethodStratify-method}
\alias{predictForCluster,lcModelLMKM-method}
\alias{coef.lcModelLMKM}
\alias{converged,lcModelLMKM-method}
\alias{postprob,lcModelLMKM-method}
\alias{getName,lcModelTwoStep-method}
\alias{getShortName,lcModelTwoStep-method}
\title{featureBased interface}
\usage{
\S4method{getName}{lcMethodTwoStep}(object)

\S4method{getShortName}{lcMethodTwoStep}(object)

\S4method{prepareData}{lcMethodTwoStep}(method, data, verbose, ...)

\S4method{fit}{lcMethodTwoStep}(method, data, envir, verbose, ...)

\S4method{getName}{lcMethodGCKM}(object)

\S4method{getShortName}{lcMethodGCKM}(object)

\S4method{compose}{lcMethodGCKM}(method, envir = NULL)

\S4method{preFit}{lcMethodGCKM}(method, data, envir, verbose)

\S4method{fit}{lcMethodGCKM}(method, data, envir, verbose, ...)

\S4method{getName}{lcMethodLMKM}(object)

\S4method{getShortName}{lcMethodLMKM}(object)

\S4method{prepareData}{lcMethodLMKM}(method, data, verbose)

\S4method{fit}{lcMethodLMKM}(method, data, envir, verbose, ...)

\S4method{getName}{lcMethodStratify}(object)

\S4method{getShortName}{lcMethodStratify}(object)

\S4method{compose}{lcMethodStratify}(method, envir = NULL, ...)

\S4method{fit}{lcMethodStratify}(method, data, envir, verbose, ...)

\S4method{predictForCluster}{lcModelLMKM}(object, newdata, cluster, what = "mu", ...)

\method{coef}{lcModelLMKM}(object, ..., cluster = NULL)

\S4method{converged}{lcModelLMKM}(object, ...)

\S4method{postprob}{lcModelLMKM}(object, ...)

\S4method{getName}{lcModelTwoStep}(object, ...)

\S4method{getShortName}{lcModelTwoStep}(object, ...)
}
\arguments{
\item{object}{The object to extract the label from.}

\item{method}{The \code{lcMethod} object.}

\item{data}{The data, as a \code{data.frame}, on which the model will be trained.}

\item{verbose}{A \link[R.utils:Verbose]{R.utils::Verbose} object indicating the level of verbosity.}

\item{...}{
  Arguments passed on to \code{\link[stats:predict.lm]{stats::predict.lm}}
  \describe{
    \item{\code{se.fit}}{A switch indicating if standard errors are required.}
    \item{\code{scale}}{Scale parameter for std.err. calculation.}
    \item{\code{df}}{Degrees of freedom for scale.}
    \item{\code{interval}}{Type of interval calculation.  Can be abbreviated.}
    \item{\code{level}}{Tolerance/confidence level.}
    \item{\code{type}}{Type of prediction (response or model term).  Can be abbreviated.}
    \item{\code{terms}}{If \code{type = "terms"}, which terms (default is all
    terms), a \code{\link{character}} vector.}
    \item{\code{na.action}}{function determining what should be done with missing
    values in \code{newdata}.  The default is to predict \code{NA}.}
    \item{\code{pred.var}}{the variance(s) for future observations to be assumed
    for prediction intervals.  See \sQuote{Details}.}
    \item{\code{weights}}{variance weights for prediction.  This can be a numeric
    vector or a one-sided model formula.  In the latter case, it is
    interpreted as an expression evaluated in \code{newdata}.}
  }}

\item{envir}{The \code{environment} in which the \code{lcMethod} should be evaluated}

\item{newdata}{Optional data frame for which to compute the model predictions. If omitted, the model training data is used.
Cluster trajectory predictions are made when ids are not specified. If the clusters are specified under the Cluster column, output is given only for the specified cluster. Otherwise, a matrix is returned with predictions for all clusters.}

\item{cluster}{The cluster name.}

\item{what}{The distributional parameter to predict. By default, the mean response 'mu' is predicted. The cluster membership predictions can be obtained by specifying what='mb'.}
}
\description{
featureBased interface
}
\seealso{
\link{lcMethodTwoStep} \link{lcMethodGCKM} \link{lcMethodLMKM}
}
\keyword{internal}
