% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodFunFEM.R, R/modelFunFEM.R
\name{interface-funFEM}
\alias{interface-funFEM}
\alias{getName,lcMethodFunFEM-method}
\alias{getShortName,lcMethodFunFEM-method}
\alias{preFit,lcMethodFunFEM-method}
\alias{fit,lcMethodFunFEM-method}
\alias{fitted.lcModelFunFEM}
\alias{predictForCluster,lcModelFunFEM-method}
\alias{postprob,lcModelFunFEM-method}
\alias{coef.lcModelFunFEM}
\alias{logLik.lcModelFunFEM}
\alias{converged,lcModelFunFEM-method}
\title{funFEM interface}
\usage{
\S4method{getName}{lcMethodFunFEM}(object)

\S4method{getShortName}{lcMethodFunFEM}(object)

\S4method{preFit}{lcMethodFunFEM}(method, data, envir, verbose, ...)

\S4method{fit}{lcMethodFunFEM}(method, data, envir, verbose, ...)

\method{fitted}{lcModelFunFEM}(object, ..., clusters = trajectoryAssignments(object))

\S4method{predictForCluster}{lcModelFunFEM}(
  object,
  newdata,
  cluster,
  what = "mu",
  approxFun = approx,
  ...
)

\S4method{postprob}{lcModelFunFEM}(object, ...)

\method{coef}{lcModelFunFEM}(object, ...)

\method{logLik}{lcModelFunFEM}(object, ...)

\S4method{converged}{lcModelFunFEM}(object, ...)
}
\arguments{
\item{object}{The object to extract the label from.}

\item{method}{The \code{lcMethod} object.}

\item{data}{The data, as a \code{data.frame}, on which the model will be trained.}

\item{envir}{The \code{environment} in which the \code{lcMethod} should be evaluated}

\item{verbose}{A \link[R.utils:Verbose]{R.utils::Verbose} object indicating the level of verbosity.}

\item{...}{Additional arguments.}

\item{clusters}{Optional cluster assignments per id. If unspecified, a \code{matrix} is returned containing the cluster-specific predictions per column.}

\item{newdata}{Optional \code{data.frame} for which to compute the model predictions. If omitted, the model training data is used.
Cluster trajectory predictions are made when ids are not specified.}

\item{cluster}{The cluster name (as \code{character}) to predict for.}

\item{what}{The distributional parameter to predict. By default, the mean response 'mu' is predicted. The cluster membership predictions can be obtained by specifying \code{what = 'mb'}.}

\item{approxFun}{Function to interpolate between measurement moments, \link[stats]{approx}() by default.}
}
\description{
funFEM interface
}
\seealso{
\link{lcMethodFunFEM} \link[funFEM]{funFEM-package}
}
\keyword{internal}
