% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodMclustLLPA.R, R/modelMclustLLPA.R
\name{interface-mclust}
\alias{interface-mclust}
\alias{getName,lcMethodMclustLLPA-method}
\alias{getShortName,lcMethodMclustLLPA-method}
\alias{prepareData,lcMethodMclustLLPA-method}
\alias{compose,lcMethodMclustLLPA-method}
\alias{fit,lcMethodMclustLLPA-method}
\alias{predict.lcModelMclustLLPA}
\alias{fitted.lcModelMclustLLPA}
\alias{postprob,lcModelMclustLLPA-method}
\alias{predictPostprob,lcModelMclustLLPA-method}
\alias{converged,lcModelMclustLLPA-method}
\title{mclust interface}
\usage{
\S4method{getName}{lcMethodMclustLLPA}(object)

\S4method{getShortName}{lcMethodMclustLLPA}(object)

\S4method{prepareData}{lcMethodMclustLLPA}(method, data, verbose, ...)

\S4method{compose}{lcMethodMclustLLPA}(method, envir = NULL)

\S4method{fit}{lcMethodMclustLLPA}(method, data, envir, verbose, ...)

\method{predict}{lcModelMclustLLPA}(object, ..., newdata = NULL, what = "mu", approxFun = approx)

\method{fitted}{lcModelMclustLLPA}(object, ..., clusters = trajectoryAssignments(object))

\S4method{postprob}{lcModelMclustLLPA}(object, ...)

\S4method{predictPostprob}{lcModelMclustLLPA}(object, newdata = NULL, ...)

\S4method{converged}{lcModelMclustLLPA}(object, ...)
}
\arguments{
\item{object}{The object to extract the label from.}

\item{method}{The \code{lcMethod} object.}

\item{data}{The data, as a \code{data.frame}, on which the model will be trained.}

\item{verbose}{A \link[R.utils:Verbose]{R.utils::Verbose} object indicating the level of verbosity.}

\item{...}{Additional arguments.}

\item{envir}{The \code{environment} in which the \code{lcMethod} should be evaluated}

\item{newdata}{Optional data frame for which to compute the model predictions. If omitted, the model training data is used.
Cluster trajectory predictions are made when ids are not specified. If the clusters are specified under the Cluster column, output is given only for the specified cluster. Otherwise, a matrix is returned with predictions for all clusters.}

\item{what}{The distributional parameter to predict. By default, the mean response 'mu' is predicted. The cluster membership predictions can be obtained by specifying what='mb'.}

\item{clusters}{Optional cluster assignments per id. If unspecified, a \code{matrix} is returned containing the cluster-specific predictions per column.}
}
\description{
mclust interface
}
\seealso{
\link{lcMethodMclustLLPA} \link[mclust]{mclust-package}
}
\keyword{internal}
