% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodMclustLLPA.R, R/modelMclustLLPA.R
\name{interface-mclust}
\alias{interface-mclust}
\alias{getName,lcMethodMclustLLPA-method}
\alias{getShortName,lcMethodMclustLLPA-method}
\alias{prepareData,lcMethodMclustLLPA-method}
\alias{compose,lcMethodMclustLLPA-method}
\alias{fit,lcMethodMclustLLPA-method}
\alias{clusterTrajectories,lcModelMclustLLPA-method}
\alias{postprob,lcModelMclustLLPA-method}
\alias{predictPostprob,lcModelMclustLLPA-method}
\alias{converged,lcModelMclustLLPA-method}
\title{mclust interface}
\usage{
\S4method{getName}{lcMethodMclustLLPA}(object)

\S4method{getShortName}{lcMethodMclustLLPA}(object)

\S4method{prepareData}{lcMethodMclustLLPA}(method, data, verbose, ...)

\S4method{compose}{lcMethodMclustLLPA}(method, envir = NULL)

\S4method{fit}{lcMethodMclustLLPA}(method, data, envir, verbose, ...)

\S4method{clusterTrajectories}{lcModelMclustLLPA}(object, at = time(object), ...)

\S4method{postprob}{lcModelMclustLLPA}(object, ...)

\S4method{predictPostprob}{lcModelMclustLLPA}(object, newdata = NULL, ...)

\S4method{converged}{lcModelMclustLLPA}(object, ...)
}
\arguments{
\item{object}{The object to extract the label from.}

\item{method}{The \code{lcMethod} object.}

\item{data}{The data, as a \code{data.frame}, on which the model will be trained.}

\item{verbose}{A \link[R.utils:Verbose]{R.utils::Verbose} object indicating the level of verbosity.}

\item{...}{Additional arguments.}

\item{envir}{The \code{environment} in which the \code{lcMethod} should be evaluated}

\item{at}{An optional vector, list or data frame of covariates at which to compute the cluster trajectory predictions.
If a vector is specified, this is assumed to be the time covariate. Otherwise, a named list or data frame must be provided.}

\item{newdata}{Optional \code{data.frame} for which to compute the model predictions. If omitted, the model training data is used.
Cluster trajectory predictions are made when ids are not specified.}
}
\description{
mclust interface
}
\seealso{
\link{lcMethodMclustLLPA} \link[mclust]{mclust-package}
}
\keyword{internal}
