% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-method.R, R/meta-fit-converged.R
\docType{class}
\name{interface-metaMethods}
\alias{interface-metaMethods}
\alias{lcMetaMethod-class}
\alias{compose,lcMetaMethod-method}
\alias{getLcMethod,lcMetaMethod-method}
\alias{getName,lcMetaMethod-method}
\alias{getShortName,lcMetaMethod-method}
\alias{idVariable,lcMetaMethod-method}
\alias{preFit,lcMetaMethod-method}
\alias{prepareData,lcMetaMethod-method}
\alias{postFit,lcMetaMethod-method}
\alias{responseVariable,lcMetaMethod-method}
\alias{timeVariable,lcMetaMethod-method}
\alias{validate,lcMetaMethod-method}
\alias{fit,lcFitConverged-method}
\alias{validate,lcFitConverged-method}
\title{lcMetaMethod abstract class}
\usage{
\S4method{compose}{lcMetaMethod}(method, envir = NULL)

\S4method{getLcMethod}{lcMetaMethod}(object, ...)

\S4method{getName}{lcMetaMethod}(object, ...)

\S4method{getShortName}{lcMetaMethod}(object, ...)

\S4method{idVariable}{lcMetaMethod}(object, ...)

\S4method{preFit}{lcMetaMethod}(method, data, envir, verbose)

\S4method{prepareData}{lcMetaMethod}(method, data, verbose)

\S4method{postFit}{lcMetaMethod}(method, data, model, envir, verbose)

\S4method{responseVariable}{lcMetaMethod}(object, ...)

\S4method{timeVariable}{lcMetaMethod}(object, ...)

\S4method{validate}{lcMetaMethod}(method, data, envir = NULL, ...)

\S4method{fit}{lcFitConverged}(method, data, envir, verbose)

\S4method{validate}{lcFitConverged}(method, data, envir = NULL, ...)
}
\description{
Virtual class for internal use. Do not use.
}
