% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodMixTVEM.R, R/modelMixTVEM.R
\name{interface-mixtvem}
\alias{interface-mixtvem}
\alias{getName,lcMethodMixTVEM-method}
\alias{getShortName,lcMethodMixTVEM-method}
\alias{preFit,lcMethodMixTVEM-method}
\alias{fit,lcMethodMixTVEM-method}
\alias{predict.lcModelMixTVEM}
\alias{postprob,lcModelMixTVEM-method}
\alias{converged,lcModelMixTVEM-method}
\alias{logLik.lcModelMixTVEM}
\alias{sigma.lcModelMixTVEM}
\alias{coef.lcModelMixTVEM}
\title{mixtvem interface}
\usage{
\S4method{getName}{lcMethodMixTVEM}(object)

\S4method{getShortName}{lcMethodMixTVEM}(object)

\S4method{preFit}{lcMethodMixTVEM}(method, data, envir, verbose, ...)

\S4method{fit}{lcMethodMixTVEM}(method, data, envir, verbose, ...)

\method{predict}{lcModelMixTVEM}(object, ..., newdata = NULL, what = "mu")

\S4method{postprob}{lcModelMixTVEM}(object, ...)

\S4method{converged}{lcModelMixTVEM}(object, ...)

\method{logLik}{lcModelMixTVEM}(object, ...)

\method{sigma}{lcModelMixTVEM}(object, ...)

\method{coef}{lcModelMixTVEM}(object, ...)
}
\arguments{
\item{object}{The object to extract the label from.}

\item{method}{An object inheriting from \code{lcMethod} with all its arguments having been evaluated and finalized.}

\item{data}{A \code{data.frame} representing the transformed training data.}

\item{verbose}{A \link[R.utils:Verbose]{R.utils::Verbose} object indicating the level of verbosity.}

\item{...}{Additional arguments.}

\item{newdata}{Optional \code{data.frame} for which to compute the model predictions. If omitted, the model training data is used.
Cluster trajectory predictions are made when ids are not specified.}

\item{what}{The distributional parameter to predict. By default, the mean response 'mu' is predicted. The cluster membership predictions can be obtained by specifying \code{what = 'mb'}.}
}
\description{
mixtvem interface
}
\seealso{
\link{lcMethodMixTVEM}
}
\keyword{internal}
