% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{latrend-approaches}
\alias{latrend-approaches}
\title{High-level approaches to longitudinal clustering}
\description{
This page provides high-level guidelines on which methods are applicable to your dataset.
Note that this is intended as a quick-start.

Recommended overview and comparison papers:
\itemize{
\item \insertCite{denteuling2021clustering}{latrend}: A tutorial and overview on methods for longitudinal clustering.
\item \insertCite{denteuling2021comparison;textual}{latrend} compared KmL, MixTVEM, GBTM, GMM, and GCKM.
\item \insertCite{twisk2012classifying;textual}{latrend} compared KmL, GCKM, LLCA, GBTM and GMM.
\item \insertCite{verboon2022clustering;textual}{latrend} compared the \emph{kml}, \emph{traj} and \emph{lcmm} packages in R.
\item \insertCite{martin2015growth;textual}{latrend} compared KmL, LCA, and GMM.
}
}
\section{Approaches}{

Disclaimer: The table below has been adapted from a pre-print of \insertCite{denteuling2021clustering}{latrend}.\tabular{llll}{
   \strong{Approach} \tab \strong{Strengths} \tab \strong{Limitations} \tab \strong{Methods} \cr
   \strong{Cross-sectional clustering} \tab \itemize{\item Suitable for large datasets \item Many available algorithms \item Non-parametric cluster trajectory representation} \tab \itemize{\item Requires time-aligned complete data \item Sensitive to measurement noise} \tab \link{lcMethodKML} \link{lcMethodMclustLLPA} \link{lcMethodMixtoolsNPRM} \cr
   \strong{Distance-based clustering} \tab \itemize{\item Suitable for medium-sized datasets \item Many possible distance metrics \item Distance matrix only needs to be computed once} \tab \itemize{\item Scales poorly with number of trajectories \item No robust cluster trajectory representation \item Some distance metrics require aligned observations} \tab \link{lcMethodDtwclust} \cr
   \strong{Feature-based clustering} \tab \itemize{\item Suitable for large datasets \item Configurable \item Features only needs to be computed once \item Compact trajectory representation} \tab \itemize{\item Generally requires intensive longitudinal data \item Sensitive to outliers} \tab \link{lcMethodFeature} \link{lcMethodAkmedoids} \link{lcMethodLMKM} \link{lcMethodGCKM} \cr
   \strong{Model-based clustering} \tab \itemize{\item Parametric cluster trajectory \item Incorporate (domain) assumptions \item Low sample size requirements} \tab \itemize{\item Computationally intensive \item Scales poorly with number of clusters \item Convergence challenges} \tab \link{lcMethodLcmmGBTM} \link{lcMethodLcmmGMM} \link{lcMethodCrimCV} \link{lcMethodFlexmix} \link{lcMethodFlexmixGBTM} \link{lcMethodFunFEM} \link{lcMethodMixAK_GLMM} \link{lcMethodMixtoolsGMM} \link{lcMethodMixTVEM} \cr
}


It is strongly encouraged to \link[=latrend-metrics]{evaluate and compare} several candidate methods in order to identify the most suitable method.
}

\references{
\insertAllCited{}
}
\seealso{
\link{latrend-methods} \link{latrend-estimation} \link{latrend-metrics}
}
