% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{latrendData}
\alias{latrendData}
\title{Synthetic longitudinal dataset comprising three classes}
\format{
A \code{data.frame} describing 250 trajectories originating from one of three classes,
each with a different cluster trajectory. Trajectories randomly deviate in intercept and slope from the reference cluster.
\describe{
\item{Id}{trajectory identifier, \code{factor}.}
\item{Time}{measurement time, \code{numeric} between 0 and 2.}
\item{Y}{observed variable, \code{numeric}.}
\item{Class}{the reference class, \code{factor}.}
}
}
\source{
This dataset was generated using \link{generateLongData}.
}
\usage{
latrendData
}
\description{
Synthetic longitudinal dataset comprising three classes
}
\seealso{
\link{generateLongData}
}
\keyword{datasets}
