% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelApprox.R
\docType{class}
\name{lcApproxModel-class}
\alias{lcApproxModel-class}
\alias{lcApproxModel}
\alias{fitted.lcApproxModel}
\alias{predict.lcApproxModel}
\title{lcApproxModel class}
\usage{
\method{fitted}{lcApproxModel}(object, ..., clusters = trajectoryAssignments(object))

\method{predict}{lcApproxModel}(object, ..., newdata = NULL, what = "mu", approxFun = approx)
}
\arguments{
\item{object}{The \code{lcModel} object.}

\item{...}{Additional arguments.}

\item{clusters}{Optional cluster assignments per id. If unspecified, a \code{matrix} is returned containing the cluster-specific predictions per column.}

\item{newdata}{Optional data frame for which to compute the model predictions. If omitted, the model training data is used.
Cluster trajectory predictions are made when ids are not specified. If the clusters are specified under the Cluster column, output is given only for the specified cluster. Otherwise, a matrix is returned with predictions for all clusters.}

\item{what}{The distributional parameter to predict. By default, the mean response 'mu' is predicted. The cluster membership predictions can be obtained by specifying what='mb'.}

\item{approxFun}{The interpolation function to use for time points not in the feature set.}
}
\description{
approx models have defined cluster trajectories at fixed moments in time, which should be interpolated
For a correct implementation, lcApproxModel requires the extending class to implement clusterTrajectories(at=NULL)
to return the fixed cluster trajectories
}
