% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodMatrix.R
\docType{class}
\name{lcMatrixMethod-class}
\alias{lcMatrixMethod-class}
\alias{getName,lcMatrixMethod-method}
\alias{getShortName,lcMatrixMethod-method}
\alias{prepareData,lcMatrixMethod-method}
\title{lcMatrixMethod}
\usage{
\S4method{getName}{lcMatrixMethod}(object, ...)

\S4method{getShortName}{lcMatrixMethod}(object, ...)

\S4method{prepareData}{lcMatrixMethod}(method, data, verbose, ...)
}
\arguments{
\item{object}{The data or model or extract the trajectories from.}

\item{...}{Not used.}

\item{method}{An object inheriting from \code{lcMethod} with all its arguments having been evaluated and finalized.}

\item{data}{A \code{data.frame} representing the transformed training data.}

\item{verbose}{A \link[R.utils:Verbose]{R.utils::Verbose} object indicating the level of verbosity.}
}
\description{
lcMatrixMethod
}
\keyword{internal}
