% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{lcModel-make}
\alias{lcModel-make}
\alias{make.trajectoryAssignments}
\alias{make.clusterIndices}
\alias{make.clusterNames}
\title{Cluster-handling functions for lcModel implementations.}
\usage{
make.trajectoryAssignments(object, clusters)

make.clusterIndices(object, clusters)

make.clusterNames(n)
}
\arguments{
\item{object}{The \code{lcModel} object.}

\item{clusters}{The unprocessed trajectory cluster assignment vector.}

\item{n}{The number of clusters.}
}
\value{
Factor cluster assignments.

A cluster assignments index vector of type \code{integer}.

A \code{character} vector length \code{n} with the cluster names.
}
\description{
Ensures a proper cluster assignments factor vector
}
\keyword{internal}
