% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model.data.lcModel}
\alias{model.data.lcModel}
\title{Extract the model data that was used for fitting}
\usage{
\method{model.data}{lcModel}(object, ...)
}
\arguments{
\item{object}{The \code{lcModel} object.}

\item{...}{Additional arguments.}
}
\value{
The \code{data.frame} that was used for fitting the \code{lcModel}.
}
\description{
Evaluates the data call in the environment that the model was trained in.
}
