% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model.data.lcModel}
\alias{model.data.lcModel}
\title{Extract the model data that was used for fitting}
\usage{
\method{model.data}{lcModel}(object, ...)
}
\arguments{
\item{object}{The \code{lcModel} object.}

\item{...}{Additional arguments.}
}
\value{
The full \code{data.frame} that was used for fitting the \code{lcModel}.
}
\description{
Evaluates the data call in the environment that the model was trained in.
}
\examples{
library(kml)
data(latrendData)
method <- lcMethodKML("Y", id = "Id", time = "Time", nClusters = 3)
kml <- latrend(method, latrendData)
model.data(kml)
}
\seealso{
\link{model.frame.lcModel} \link{time.lcModel}
}
