% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{nClusters}
\alias{nClusters}
\title{Number of clusters}
\usage{
nClusters(object)
}
\arguments{
\item{object}{The \code{lcModel} object.}
}
\value{
An \code{integer} with the number of clusters identified by the \code{lcModel}.
}
\description{
Get the number of clusters estimated by the given \code{lcModel} object.
}
\examples{
data(latrendData)
method <- lcMethodRandom("Y", id = "Id", time = "Time", nClusters = 3)
model <- latrend(method, latrendData)
nClusters(model) # 3
}
\seealso{
\link{nIds} \link{nobs}
}
