% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{postprob}
\alias{postprob}
\alias{postprob,lcModel-method}
\title{Posterior probability per fitted id}
\usage{
\S4method{postprob}{lcModel}(object, ...)
}
\arguments{
\item{object}{The \code{lcModel}.}

\item{...}{Additional arguments.}
}
\description{
Posterior probability per fitted id
}
\examples{
data(latrendData)
model <- latrend(lcMethodLcmmGMM(fixed = Y ~ Time, mixture = ~ Time,
   id = "Id", time = "Time"), data = latrendData)
postprob(model)
}
\seealso{
Other model-specific methods: 
\code{\link{clusterTrajectories}()},
\code{\link{coef.lcModel}()},
\code{\link{converged}()},
\code{\link{deviance.lcModel}()},
\code{\link{df.residual.lcModel}()},
\code{\link{fitted.lcModel}()},
\code{\link{lcModel-class}},
\code{\link{logLik.lcModel}()},
\code{\link{model.frame.lcModel}()},
\code{\link{nobs.lcModel}()},
\code{\link{predict.lcModel}()},
\code{\link{predictAssignments}()},
\code{\link{predictForCluster}()},
\code{\link{predictPostprob}()},
\code{\link{residuals.lcModel}()},
\code{\link{sigma.lcModel}()},
\code{\link{time.lcModel}()},
\code{\link{trajectories}()}
}
\concept{model-specific methods}
