% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{predictPostprob}
\alias{predictPostprob}
\alias{predictPostprob,lcModel-method}
\title{lcModel posterior probability prediction}
\usage{
\S4method{predictPostprob}{lcModel}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{The \code{lcModel} to predict the posterior probabilities with.}

\item{newdata}{Optional data frame for which to compute the posterior probability. If omitted, the model training data is used.}

\item{...}{Additional arguments.}
}
\value{
A \code{matrix} indicating the posterior probability per trajectory per measurement on each row, for each cluster (the columns).
}
\description{
lcModel posterior probability prediction
}
\details{
The default implementation returns a uniform probability matrix.
}
\seealso{
Other model-specific methods: 
\code{\link{clusterTrajectories}()},
\code{\link{coef.lcModel}()},
\code{\link{converged}()},
\code{\link{deviance.lcModel}()},
\code{\link{df.residual.lcModel}()},
\code{\link{fitted.lcModel}()},
\code{\link{lcModel-class}},
\code{\link{logLik.lcModel}()},
\code{\link{model.frame.lcModel}()},
\code{\link{nobs.lcModel}()},
\code{\link{postprob}()},
\code{\link{predict.lcModel}()},
\code{\link{predictAssignments}()},
\code{\link{predictForCluster}()},
\code{\link{residuals.lcModel}()},
\code{\link{sigma.lcModel}()},
\code{\link{time.lcModel}()},
\code{\link{trajectories}()}
}
\concept{model-specific methods}
