% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{print.lcModels}
\alias{print.lcModels}
\title{Print lcModels list concisely}
\usage{
\method{print}{lcModels}(
  x,
  ...,
  summary = FALSE,
  excludeShared = !getOption("latrend.printSharedModelArgs")
)
}
\arguments{
\item{x}{The \code{lcModels} object.}

\item{...}{Not used.}

\item{summary}{Whether to print the complete summary per model. This may be slow for long lists!}

\item{excludeShared}{Whether to exclude model arguments which are identical across all models.}
}
\description{
Print lcModels list concisely
}
\seealso{
Other lcModel list functions: 
\code{\link{as.lcModels}()},
\code{\link{lcModels}},
\code{\link{subset.lcModels}()}
}
\concept{lcModel list functions}
