% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{qqPlot}
\alias{qqPlot}
\alias{qqPlot,lcModel-method}
\title{Quantile-quantile plot}
\usage{
\S4method{qqPlot}{lcModel}(object, byCluster = FALSE, ...)
}
\arguments{
\item{object}{The model.}

\item{byCluster}{Whether to plot the Q-Q line per cluster}

\item{...}{Other arguments passed to qqplotr::geom_qq_band, qqplotr::stat_qq_line, and qqplotr::stat_qq_point.}
}
\description{
Quantile-quantile plot
}
\seealso{
\link{residuals.lcModel} \link{metric} \link{plotClusterTrajectories}
}
