% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{strip}
\alias{strip}
\alias{strip,lcModel-method}
\title{Strip a lcModel for serialization}
\usage{
\S4method{strip}{lcModel}(object, ...)
}
\arguments{
\item{object}{The \code{lcModel}.}

\item{...}{Additional arguments.}
}
\description{
Removes associated environments from any of the slots.
}
