% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R, R/model.R
\name{strip}
\alias{strip}
\alias{strip,lcMethod-method}
\alias{strip,lcModel-method}
\title{Strip a lcModel for serialization}
\usage{
\S4method{strip}{lcMethod}(object, ...)

\S4method{strip}{lcModel}(object, ...)
}
\arguments{
\item{object}{The \code{lcModel}.}

\item{...}{Additional arguments.}
}
\description{
Removes associated environments from any of the arguments. This is typically the case for arguments of type \code{formula}.

Strip a lcModel of non-essential variables and environments in order to reduce the model size for serialization.
}
