% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R, R/model.R
\name{strip}
\alias{strip}
\alias{strip,lcMethod-method}
\alias{strip,lcModel-method}
\title{Strip a lcModel for serialization}
\usage{
\S4method{strip}{lcMethod}(object, ..., classes = "formula")

\S4method{strip}{ANY}(object, ..., classes = "formula")

\S4method{strip}{lcModel}(object, ..., classes = "formula")
}
\arguments{
\item{object}{The \code{lcModel}.}

\item{...}{Additional arguments.}

\item{classes}{The object classes for which to remove their assigned environment. By default, only environments from \code{formula} are removed.}
}
\description{
Strip a lcModel of non-essential variables and environments in order to reduce the model size for serialization.
}
