% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R, R/model.R
\name{strip}
\alias{strip}
\alias{strip,lcMethod-method}
\alias{strip,lcModel-method}
\title{Strip a lcModel for serialization}
\usage{
\S4method{strip}{lcMethod}(object, ..., classes = "formula")

\S4method{strip}{ANY}(object, ..., classes = "formula")

\S4method{strip}{lcModel}(object, ..., classes = "formula")
}
\arguments{
\item{object}{The \code{lcModel}.}

\item{...}{Additional arguments.}

\item{classes}{The object classes for which to remove their assigned environment. By default, only environments from \code{formula} are removed.}
}
\value{
An \code{lcModel} object of the same type as the \code{object} argument.
}
\description{
Strip a lcModel of non-essential variables and environments in order to reduce the model size for serialization.
}
\section{Implementation}{

Classes extending \code{lcModel} can override this method to remove additional non-essentials.
\preformatted{
setMethod("strip", "lcModelExt", function(object, ..., classes = "formula") {
  object <- callNextMethod()
  # further process the object
  return (object)
})
}
}

\examples{
library(lcmm)
data(latrendData)
model <- latrend(lcMethodLcmmGMM(fixed = Y ~ Time, mixture = ~ Time,
   id = "Id", time = "Time"), data = latrendData)
strip(model)
}
