% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test.latrend}
\alias{test.latrend}
\title{Test the implementation of an lcMethod and associated lcModel subclasses}
\usage{
test.latrend(
  class = "lcMethodKML",
  instantiator = NULL,
  args = list(),
  tests = c("method", "basic", "fitted", "predict", "cluster-single", "cluster-three"),
  maxFails = 5L,
  errorOnFail = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{class}{The name of the \code{lcMethod} subclass to test.
The class should inherit from \code{lcMethod}.}

\item{instantiator}{A \code{function} with signature \verb{(id, time, response, ...)},
returning an object inheriting from the \code{lcMethod} specified by the \code{class} argument.}

\item{args}{Other arguments passed to the instantiator function.}

\item{tests}{A \code{character} vector indicating the type of tests to run, as defined in the \verb{*.Rraw} files inside the \verb{/test/} folder.}

\item{errorOnFail}{Whether to throw the test errors as an error. This is always enabled while running package tests.}

\item{verbose}{Whether the output testing results. This is always disabled while running package tests.}
}
\description{
Test a \code{lcMethod} subclass implementation and its resulting \code{lcModel} implementation.
}
\note{
This is an experimental function that is subject to large changes in the future.
}
\examples{
test.latrend("lcMethodRandom")
}
