% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R, R/model.R
\name{trajectoryAssignments}
\alias{trajectoryAssignments}
\alias{trajectoryAssignments,matrix-method}
\alias{trajectoryAssignments,lcModel-method}
\title{Get the cluster membership of each trajectory}
\usage{
\S4method{trajectoryAssignments}{matrix}(
  object,
  strategy = which.max,
  clusterNames = colnames(object),
  ...
)

\S4method{trajectoryAssignments}{lcModel}(object, strategy = which.max, ...)
}
\arguments{
\item{object}{The \code{lcModel} to obtain the cluster assignments from.}

\item{strategy}{A function returning the cluster index based on the given vector of membership probabilities. By default, ids are assigned to the cluster with the highest probability.}

\item{clusterNames}{Optional \verb{character vector} with the cluster names. If \code{clusterNames = NULL}, \code{\link[=make.clusterNames]{make.clusterNames()}} is used.}

\item{...}{Any additional arguments passed to the strategy function.}
}
\value{
A \code{factor} indicating the cluster membership for each trajectory, of length \code{nIds(object)}.
}
\description{
Classify the fitted trajectories based on the posterior probabilities computed by \code{\link[=postprob]{postprob()}}, according to a given classification strategy.

By default, trajectories are assigned based on the highest posterior probability using \code{\link[=which.max]{which.max()}}.
In cases where identical probabilities are expected between clusters, it is preferable to use \link[nnet]{which.is.max} instead, as this function breaks ties at random.
Another strategy to consider is the function \code{\link[=which.weight]{which.weight()}}, which enables weighted sampling of cluster assignments based on the trajectory-specific probabilities.
}
\examples{
library(kml)
data(latrendData)
model <- latrend(method = lcMethodKML("Y", id = "Id", time = "Time"), latrendData)
trajectoryAssignments(model)

# assign trajectories at random using weighted sampling
trajectoryAssignments(model, strategy = which.weight)
}
\seealso{
\link{postprob} \link{clusterSizes} \link{predictAssignments}
}
